% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_score_spacers.R
\name{add_context}
\alias{add_context}
\title{Add [-4, +3] context}
\usage{
add_context(spacers, bsgenome, verbose = TRUE)
}
\arguments{
\item{spacers}{\code{\link[GenomicRanges]{GRanges-class}}: spacer ranges}

\item{bsgenome}{\code{\link[BSgenome]{BSgenome-class}}}

\item{verbose}{logical(1)}
}
\value{
character vector
}
\description{
Add context for Doench2016 scoring
}
\examples{
# PE example
#-----------
    require(magrittr)
    bsgenome <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38  
    gr <- char_to_granges(c(PRNP = 'chr20:4699600:+',             # snp
                            HBB  = 'chr11:5227002:-',             # snp
                            HEXA = 'chr15:72346580-72346583:-',   # del
                            CFTR = 'chr7:117559593-117559595:+'), # ins
                          bsgenome)
    spacers <- find_spacers(extend_for_pe(gr), bsgenome, complement=FALSE)
    (add_context(spacers, bsgenome))

# TFBS example
#-------------
    bedfile <- system.file('extdata/SRF.bed', package = 'multicrispr')
    bsgenome <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
    targets  <- extend(bed_to_granges(bedfile, 'mm10'))
    spacers    <- find_spacers(targets, bsgenome)
    (spacers \%<>\% add_context(bsgenome))
}
