% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_filter_specific.R
\name{add_genome_counts}
\alias{add_genome_counts}
\title{Add genome counts}
\usage{
add_genome_counts(
  spacers,
  indexedgenome = default_indexedgenome(spacers),
  mismatches = 2,
  outdir = default_outdir(),
  pam = "NGG",
  verbose = TRUE
)
}
\arguments{
\item{spacers}{spacer \code{\link[GenomicRanges]{GRanges-class}}}

\item{indexedgenome}{dir with bowtie-indexed genome}

\item{mismatches}{number (default 2): max number of mismatches to consider}

\item{outdir}{dir where output is written to}

\item{pam}{string (default 'NGG') pam pattern to expand}

\item{verbose}{TRUE (default) or FALSE}
}
\description{
Count spacer matches to genome and add to GRanges
}
\details{
Expands iupac amgiguities in the pam sequence.
Matches all resulting sequences against (indexes) target and genome.
Adds match counts to GRanges object, and then returns it.
}
\examples{
# PE example
#-----------
 bsgenome <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38  
 gr <- char_to_granges(c(PRNP = 'chr20:4699600:+',             # snp
                         HBB  = 'chr11:5227002:-',             # snp
                         HEXA = 'chr15:72346580-72346583:-',   # del
                         CFTR = 'chr7:117559593-117559595:+'), # ins
                       bsgenome)
 spacers <- find_pe_spacers(gr, bsgenome)
 indexedgenome <- '~/.multicrispr/bowtie/BSgenome.Hsapiens.UCSC.hg38'
 # index_genome(bsgenome, indexedgenome)
 # add_genome_counts(spacers, indexedgenome, mismatches=1)
    
# TFBS example
#-------------
 bsgenome <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
 bedfile  <- system.file('extdata/SRF.bed', package = 'multicrispr')
 targets <- extend(bed_to_granges(bedfile, genome = 'mm10'))
 spacers <- find_spacers(targets, bsgenome)
 indexedgenome <- '~/.multicrispr/bowtie/BSgenome.Mmusculus.UCSC.mm10'
 # index_genome(bsgenome, indexedgenome)
 # add_genome_counts(spacers, indexedgenome)
 # add_genome_counts(spacers, indexedgenome, mismatches=3)
}
\seealso{
\code{\link{index_genome}}, \code{\link{index_targets}}
}
