% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_add_offtargets.R
\name{add_specificity}
\alias{add_specificity}
\alias{add_offtargets}
\alias{filter_offtargets}
\title{Add offtarget counts}
\usage{
add_specificity(...)

add_offtargets(
  spacers,
  bsgenome,
  targets = NULL,
  mismatches = 2,
  pam = "NGG",
  method = c("bowtie", "pdict")[1],
  outdir = OUTDIR,
  indexedgenomesdir = INDEXEDGENOMESDIR,
  verbose = TRUE,
  plot = TRUE,
  ...
)

filter_offtargets(
  spacers,
  bsgenome,
  targets = NULL,
  by = "targetname",
  mismatches = 2,
  pam = "NGG",
  outdir = OUTDIR,
  indexedgenomesdir = INDEXEDGENOMESDIR,
  verbose = TRUE,
  plot = TRUE,
  ...
)
}
\arguments{
\item{...}{passed to plot_intervals}

\item{spacers}{spacer \code{\link[GenomicRanges]{GRanges-class}}}

\item{bsgenome}{\code{\link[BSgenome]{BSgenome-class}}}

\item{targets}{NULL (default) or target 
\code{\link[GenomicRanges]{GRanges-class}}}

\item{mismatches}{number (default 2): max number of mismatches to consider}

\item{pam}{string (default 'NGG'): pam sequence}

\item{method}{'bowtie' or 'pdict'}

\item{outdir}{directory where output is written to}

\item{indexedgenomesdir}{string: dir with indexed genomes}

\item{verbose}{TRUE (default) or FALSE}

\item{plot}{TRUE (default) or FALSE}

\item{by}{string (default "targetname"): filter by this mcol}
}
\value{
filtered spacer \code{\link[GenomicRanges]{GRanges-class}}
}
\description{
Add mcols 'off' (offtargets), 'off0' (0-mismatch offtargets), etc.
}
\details{
For NULL targets: offtarget counts = (G0- 1) +  G1     + ...
Non-NULL targets: offtarget counts = (G0-T0) + (G1-T1) + ...
}
\examples{
# PE example
#-----------
 require(magrittr)
 bsgenome <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38  
 gr <- char_to_granges(c(PRNP = 'chr20:4699600:+',             # snp
                         HBB  = 'chr11:5227002:-',             # snp
                         HEXA = 'chr15:72346580-72346583:-',   # del
                         CFTR = 'chr7:117559593-117559595:+'), # ins
                       bsgenome)
 spacers <- find_primespacers(gr, bsgenome)
 # index_genome(bsgenome)
 add_offtargets(spacers, bsgenome, mismatches=0)
 filter_offtargets(spacers, bsgenome, mismatches=0)
 
# TFBS example
#-------------
 bsgenome <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
 bedfile  <- system.file('extdata/SRF.bed', package = 'multicrispr')
 targets <- extend(bed_to_granges(bedfile, genome = 'mm10'))
 spacers <- find_spacers(targets, bsgenome)
 spacers \%<>\% extract(1:500)
 # index_genome(bsgenome)
 # add_offtargets(spacers, bsgenome)    # off = G - 1
 # add_offtargets(spacers, bsgenome, targets)  # off = G - T
}
