% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_score_cas9s.R
\name{add_ontargetscores}
\alias{add_ontargetscores}
\title{Add cas9 ontarget scores}
\usage{
add_ontargetscores(cas9dt, bsgenome)
}
\arguments{
\item{cas9dt}{data.table(chr, cas9start, cas9end, strand)}

\item{bsgenome}{BSGenome object, e.g. BSgenome.Mmusculus.UCSC.mm10::Mmusculus}
}
\value{
data.table(chr, cas9start, cas9end, strand)
}
\description{
Add cas9 ontarget scores
}
\examples{
require(magrittr)
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
bedfile <- system.file('extdata/SRF_sites.bed', package = 'crisprapex')
cas9dt <- read_bed(bedfile) \%>\% slop_fourways() \%>\% find_cas9s(bsgenome)
cas9dt \%>\% add_contextseq()
}
