% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_find_cas9s.R
\name{add_sequence}
\alias{add_sequence}
\title{Add sequence to granges mcols}
\usage{
add_sequence(granges)
}
\arguments{
\item{granges}{GenomicRanges::GRanges}
}
\value{
GenomicRanges::GRanges
}
\description{
Add sequence to granges mcols
}
