% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_xxx_to_granges.R
\name{annotate_granges}
\alias{annotate_granges}
\title{Get BSgenome}
\usage{
annotate_granges(gr, txdb)
}
\arguments{
\item{gr}{\code{\link[GenomicRanges]{GRanges-class}}
Annotate GRanges}

\item{txdb}{\code{\link[GenomicFeatures]{TxDb-class}}}

\item{gr}{\code{\link[GenomicRanges]{GRanges-class}}}
}
\value{
\code{\link[GenomicRanges]{GRanges-class}}
}
\description{
Get BSgenome
}
\examples{
bedfile <- system.file('extdata/SRF.bed', package = 'multicrispr')
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
gr <- bed_to_granges(bedfile, bsgenome, plot = FALSE)
txdbname <- 'TxDb.Mmusculus.UCSC.mm10.knownGene'
txdb <- utils::getFromNamespace(txdbname, txdbname)
annotate_granges(gr, txdb)
}
\seealso{
\code{\link{genes_to_granges}} (inverse operation)
}
