% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_read_assert.R
\name{as.data.table.GRanges}
\alias{as.data.table.GRanges}
\title{Convert GenomicRanges into data.table}
\usage{
\method{as.data.table}{GRanges}(x)
}
\arguments{
\item{x}{}
}
\value{
data.table(chr, start, end, end, strand)
}
\description{
Convert GenomicRanges into data.table
}
\examples{
require(magrittr)
x <- read_bed(system.file('extdata/SRF_sites.bed', package = 'crisprapex'))
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
x \%>\% as.granges(bsgenome = bsgenome) \%>\% as.data.table()
}
