% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_read_assert.R
\name{as.granges}
\alias{as.granges}
\title{Convert data.table into GenomicRanges::GRanges}
\usage{
as.granges(dt, bsgenome)
}
\arguments{
\item{bsgenome}{BSgenome, e.g. BSgenome.Mmusculus.UCSC.mm10::Mmusculus}

\item{x}{data.table}
}
\value{
GenomicRanges::GRanges
}
\description{
Convert data.table into GenomicRanges::GRanges
}
\examples{
require(magrittr)
bedfile <- system.file('extdata/SRF_sites.bed', package = 'crisprapex')
x <- read_bed(bedfile)
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
x \%>\% as.granges(bsgenome = bsgenome)
}
