% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_read_assert.R
\name{assert_is_granges_datatable}
\alias{assert_is_granges_datatable}
\title{Assert that object is a genomic ranges datatable}
\usage{
assert_is_granges_datatable(x, bsgenome)
}
\arguments{
\item{x}{data.table}

\item{bsgenome}{BSgenome, e.g. BSgenome.Mmusculus.UCSC.mm10::Mmusculus}
}
\value{
invisible(x)
}
\description{
Assert that object is a genomic ranges datatable
}
\examples{
x <- read_bed(system.file('extdata/SRF_sites.bed', package = 'crisprapex'))
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
assert_is_granges_datatable(x, bsgenome)
}
