% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_score_cas9s.R
\name{calc_ontargetscore}
\alias{calc_ontargetscore}
\title{Calculate azimuth ontargetscores}
\usage{
calc_ontargetscore(contextseqs)
}
\arguments{
\item{contextseqs}{character vector}
}
\value{
numeric vector
}
\description{
Calculate azimuth ontargetscores
}
\examples{
if (reticulate::py_module_available('azimuth')){
    contextseqs <- c('TGCCCTTATATTGTCTCCAGCAGAAGGTGT', 
                    'CCAAATATTGTCAAGTTGACAACCAGGAAT')
    calc_ontargetscore(contextseqs)
}
}
