% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utilities.R
\name{canonicalseqlevels}
\alias{canonicalseqlevels}
\title{Get canonical seqlevels}
\usage{
canonicalseqlevels(gr)
}
\arguments{
\item{gr}{\code{\link[GenomicRanges]{GRanges-class}}}
}
\value{
character vector
}
\description{
Get canonical seqlevels
}
\examples{

# TxDb mouse
    txdbname <- 'TxDb.Mmusculus.UCSC.mm10.knownGene'
    txdb <- utils::getFromNamespace(txdbname, txdbname)
    gr <- GenomicFeatures::genes(txdb)
    GenomeInfoDb::seqlevels(gr)
    canonicalseqlevels(gr)

# TxDb Human
    txdbname <- 'TxDb.Hsapiens.UCSC.hg38.knownGene'
    txdb <- utils::getFromNamespace(txdbname, txdbname)
    gr <- GenomicFeatures::genes(txdb)
    GenomeInfoDb::seqlevels(gr)
    canonicalseqlevels(gr)
    
# EnsDb mouse
    ensdb <- EnsDb.Mmusculus.v98()
    gr <- ensembldb::genes(ensdb)
    GenomeInfoDb::seqlevels(gr)
    canonicalseqlevels(gr)
    
# EnsDb human
    ensdb <- EnsDb.Hsapiens.v98()
    gr <- ensembldb::genes(ensdb)
    GenomeInfoDb::seqlevels(gr)
    canonicalseqlevels(gr)
    
# Bedfile mouse
    bedfile <- system.file('extdata/SRF.bed', package = 'multicrispr')
    gr <- read_bed(bedfile, 'mm10', plot = FALSE)
    GenomeInfoDb::seqlevels(gr)
    canonicalseqlevels(gr)
    
}
\seealso{
\code{\link[GenomeInfoDb]{seqinfo}}
}
