% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_find_cas9s.R
\name{chromosomes}
\alias{chromosomes}
\alias{canonicalchr}
\title{Get (canonical) chromosomes}
\usage{
chromosomes(gr)

canonicalchr(gr)
}
\arguments{
\item{gr}{\code{\link[GenomicRanges]{GRanges-class}}}
}
\value{
character vector
}
\description{
Get (canonical) chromosomes
}
\examples{
bedfile <- system.file('extdata/SRF.bed', package = 'multicrispr')
gr <- read_bed(bedfile, 'mm10')
chromosomes(gr)
canonicalchr(gr)
}
