% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_find_cas9s.R
\name{complement}
\alias{complement}
\title{Complement}
\usage{
complement(gr, plot = TRUE, verbose = TRUE)
}
\arguments{
\item{gr}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{plot}{logical(1)}

\item{verbose}{logical(1)}
}
\value{
\code{\link[GenomicRanges]{GRanges-class}}, twice as long as input
}
\description{
Adds inverse strand for each range
}
\examples{
require(magrittr)
bedfile <- system.file('extdata/SRF.bed', package = 'multicrispr')
gr <- read_bed(bedfile, 'mm10', plot = FALSE)
complement(gr)
}
