% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_score_cas9s.R
\name{contextify_start}
\alias{contextify_start}
\alias{contextify_end}
\title{Contextify start/stop}
\usage{
contextify_start(chr, start, strand, bsgenome)

contextify_end(chr, end, strand, bsgenome)
}
\arguments{
\item{chr}{character vector: chromosome values}

\item{start}{numeric vector: start positions}

\item{strand}{character vector: '+' or '-' values}

\item{bsgenome}{BSGenome, e.g. BSgenome.Mmusculus.UCSC.mm10::Mmusculus}

\item{end}{numeric vector: end positions}
}
\value{
numeric vector
}
\description{
Contextify start/stop
}
\examples{
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
contextify_start(c('chr1', 'chr1'), c(200, 200),  c('+', '-'), bsgenome)
contextify_end(  c('chr1', 'chr1'), c(200, 200),  c('+', '-'), bsgenome)
}
