% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_score_cas9s.R
\name{contextseqs}
\alias{contextseqs}
\title{Get contextseqs}
\usage{
contextseqs(cas9ranges, bsgenome)
}
\arguments{
\item{cas9ranges}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{bsgenome}{\code{\link[BSgenome]{BSgenome-class}}}
}
\value{
character vector
}
\description{
Get [-3, +3] contextseqs for given cas9ranges
}
\examples{
require(magrittr)
bedfile <- system.file('extdata/SRF.bed', package = 'multicrispr')
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
cas9ranges  <-  bed_to_granges(bedfile, bsgenome) \%>\% 
                double_flank() \%>\% 
                find_cas9s()
cas9ranges[1:3] \%>\% seqs()
cas9ranges[1:3] \%>\% contextseqs()
}
