% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_find_cas9s.R
\name{count_genome_matches}
\alias{count_genome_matches}
\title{Count genome matches}
\usage{
count_genome_matches(cas9seqs, bsgenome, mismatch,
  chromosomes = GenomeInfoDb::seqnames(bsgenome), verbose = TRUE)
}
\arguments{
\item{cas9seqs}{character() or \code{\link[Biostrings]{XStringSet-class}}}

\item{bsgenome}{\code{\link[BSgenome]{BSgenome-class}}}

\item{mismatch}{number: number of allowed mismatches}

\item{chromosomes}{character vector}

\item{verbose}{logical(1)}
}
\value{
numeric(length(cas9seqs))
}
\description{
Count genome matches
}
\examples{

# Read target ranges
require(magrittr)
bedfile <- system.file('extdata/SRF.bed', package = 'multicrispr')
targetranges <- read_bed(bedfile, 'mm10')
targetranges \%<>\% extend() \%>\% extract(1:10)

# Find cas9ranges
cas9ranges <- find_cas9s(targetranges)

# Count genome matches
bsgenome <- get_bsgenome(cas9ranges)
count_genome_matches(
    cas9seqs    = getSeq(bsgenome, cas9ranges),
    bsgenome    = bsgenome, 
    mismatch    = 0, 
    chromosomes = 'chrY',
    verbose     = TRUE)
}
\seealso{
\code{\link{count_genome_matches}}, \code{\link{BSgenome-utils}}
}
