% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_alignment_offtargetanalysis.R
\name{count_matches}
\alias{count_matches}
\title{Count matches}
\usage{
count_matches(
  seqs,
  targetdir = "~/.multicrispr/bowtie/target/target",
  genomedir = "~/.multicrispr/bowtie/genome/genome",
  outdir = "~/.multicrispr/bowtie",
  verbose = TRUE
)
}
\arguments{
\item{seqs}{sequences to match to targets and genome}

\item{targetdir}{string: dir with indexed target seqs (as produced by \code{\link{index_targets}})}

\item{genomedir}{string: dir with indexed genome seqs (as produced by \code{\link{index_genome}}}

\item{outdir}{string: dir to which to write output of mapping}

\item{verbose}{TRUE (default) or FALSE}
}
\description{
Count matches
}
\examples{
# TFBS example
#-------------
# Read genome and index
    require(magrittr)
    bsgenome <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
    # genomedir <- index_genome(bsgenome) # takes a few hours - once
    genomedir <- '~/.multicrispr/bowtie/genome/BSgenome.Mmusculus.UCSC.mm10'
     
# Read targets and index them
    bedfile  <- system.file('extdata/SRF.bed', package = 'multicrispr')
    targets0 <- bed_to_granges(bedfile, genome = 'mm10')
    targets  <- double_flank(targets0)
    targetdir <- index_targets(targets, bsgenome)
    
# Find spacers and match
    spacers <- find_spacers(targets, bsgenome)
    seqs <- unique(spacers$spacer)
    outdir <- '~/.multicrispr/bowtie'
    count_matches(seqs, targetdir, genomedir, outdir)
}
