% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_scan_genome.R
\name{count_ontargets}
\alias{count_ontargets}
\title{Count ontargets}
\usage{
count_ontargets(cas9seqs, targetseqs, bsgenome, mismatch, verbose = TRUE)
}
\arguments{
\item{targetseqs}{character}

\item{bsgenome}{BSgenome, e.g. BSgenome.Mmusculus.UCSC.mm10::Mmusculus}

\item{mismatch}{numeric(1)}

\item{verbose}{logical(1)}

\item{cas9ranges}{data.table(chr, start, end, strand)}
}
\value{
data.table
}
\description{
Count ontargets
}
\examples{
require(magrittr)
bsgenome <-  BSgenome.Mmusculus.UCSC.mm10::Mmusculus
bedfile <- system.file('extdata/SRF_sites.bed', package = 'crisprapex')

targetdt <- bedfile \%>\% read_bed() \%>\% flank_fourways(bsgenome)
cas9dt   <- bedfile \%>\% find_flankonly_cas9s(bsgenome)
cas9seqs   <- cas9dt$cas9seq \%>\% head(10)

cas9dt[cas9seqs, on = 'cas9seq'][, .N, by = 'cas9seq'][['N']]
count_ontargets(cas9seqs, targetseqs, bsgenome, mismatch=0)
cas9dt[cas9seqs[5], on='cas9seq'][, .N, by = 'cas9seq']
count_ontargets(cas9seqs[5], targetseqs, bsgenome, mismatch=0)

count_ontargets(cas9seqs, targetseqs, bsgenome, mismatch=1)
count_ontargets(cas9seqs, targetseqs, bsgenome, mismatch=2)
}
