% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_add_offtargets.R
\name{count_spacer_matches}
\alias{count_spacer_matches}
\title{Count spacer matches}
\usage{
count_spacer_matches(
  spacers,
  reference,
  mismatches = 2,
  pam = "NGG",
  method = c("bowtie", "pdict")[1],
  norc,
  outdir = OUTDIR,
  verbose = TRUE
)
}
\arguments{
\item{spacers}{spacer \code{\link[GenomicRanges]{GRanges-class}}}

\item{reference}{either directory (with Bowtie index) or 
BSgenome (when method = 'pdict')}

\item{mismatches}{number (default 2): max number of mismatches to consider}

\item{pam}{string (default 'NGG') pam pattern to expand}

\item{method}{'bowtie' or 'pdict'}

\item{norc}{TRUE or FALSE: whether to run bowtie also with revcompls
Generally TRUE for genome and FALSE for target matches, 
because target ranges generally include both strands.}

\item{outdir}{string: file where to output bowtie results}

\item{verbose}{TRUE (default) or FALSE}
}
\value{
data.table
}
\description{
Count spacer matches to reference
}
\details{
Expands iupac amgiguities in the pam sequence.
Matches all resulting sequences against (indexes) target and genome.
Adds match counts to GRanges object, and then returns it.
}
\examples{
# PE example
#-----------
 require(magrittr)
 bsgenome <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38  
 gr <- char_to_granges(c(PRNP = 'chr20:4699600:+',             # snp
                         HBB  = 'chr11:5227002:-',             # snp
                         HEXA = 'chr15:72346580-72346583:-',   # del
                         CFTR = 'chr7:117559593-117559595:+'), # ins
                       bsgenome)
 spacers <- find_primespacers(gr, bsgenome)
 # reference <- genome_dir(indexedgenomesdir = INDEXEDGENOMESDIR, bsgenome)
 # count_spacer_matches(spacers, reference, norc=TRUE, mismatches = 1)
 
# TFBS example
#-------------
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
bedfile  <- system.file('extdata/SRF.bed', package = 'multicrispr')
targets <- extend(bed_to_granges(bedfile, genome = 'mm10'))
reference <- index_targets(targets, bsgenome)
spacers <- find_spacers(targets, bsgenome)
count_spacer_matches(spacers, reference, norc=FALSE, mismatches = 1)
}
\seealso{
\code{\link{index_genome}}, \code{\link{index_targets}}
}
