% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_filter_offtargetfree_sites.R
\name{count_target_matches}
\alias{count_target_matches}
\alias{count_genome_matches}
\title{Count target/genome matches}
\usage{
count_target_matches(spacerseqs, targetseqs, mismatch, verbose = TRUE)

count_genome_matches(
  spacerseqs,
  bsgenome,
  mismatch = 0,
  pam = "NGG",
  include = GenomeInfoDb::standardChromosomes(bsgenome),
  verbose = TRUE,
  joint = FALSE
)
}
\arguments{
\item{spacerseqs}{character vector or 
\code{\link[Biostrings]{XStringSet-class}}}

\item{targetseqs}{character vector or 
\code{\link[Biostrings]{XStringSet-class}}}

\item{mismatch}{number: number of allowed mismatches}

\item{verbose}{TRUE (default) or FALSE}

\item{bsgenome}{\code{\link[BSgenome]{BSgenome-class}}}

\item{pam}{string (default 'NGG'): sequence following spacerseqs}

\item{include}{character vector: offtarget analysis chromosomes}
}
\value{
numeric(length(spacerseqs))
}
\description{
Count target/genome matches
}
\examples{
# PE example
#-----------
    require(magrittr)
    bsgenome <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38  
    gr <- GenomicRanges::GRanges(
              seqnames = c(PRNP = 'chr20:4699600',             # snp
                           HBB  = 'chr11:5227002',             # snp
                           HEXA = 'chr15:72346580-72346583',   # del
                           CFTR = 'chr7:117559593-117559595'), # ins
              strand   = c(PRNP = '+', HBB = '-', HEXA = '-', CFTR = '+'), 
              seqinfo  = BSgenome::seqinfo(bsgenome))
    spacers <- find_spacers(extend_for_pe(gr), bsgenome)
    spacerseqs <- spacers$spacer
    # count_genome_matches(spacerseqs, bsgenome, mismatch=0) # 6 mins
    # count_genome_matches(spacerseqs, bsgenome, mismatch=1) # 32 mins
    # count_genome_matches(spacerseqs, bsgenome, mismatch=2)
    # count_genome_matches(spacerseqs, bsgenome, mismatch=3)
    
# TFBS example
#-------------
    require(magrittr)
    bedfile <- system.file('extdata/SRF.bed', package = 'multicrispr')
    targets <- extend(bed_to_granges(bedfile, 'mm10'))
    bsgenome <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
    spacers <- find_spacers(targets, bsgenome)
    spacerseqs <- spacers$spacer
    count_genome_matches(spacerseqs, bsgenome, mismatch = 0, include = 'chrY')

# Count matches
    spacerseqs <- sites$seq[1:10]
    targetseqs <- targets$seq
    count_target_matches(spacerseqs, targetseqs, 0)
    count_genome_matches(spacerseqs, bsgenome,    0, include = 'chrY')
}
