% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_manipulate_ranges.R
\name{double_flank}
\alias{double_flank}
\title{Double flank}
\usage{
double_flank(
  gr,
  leftstart = -200,
  leftend = -1,
  rightstart = 1,
  rightend = 200,
  bsgenome = NULL,
  verbose = TRUE,
  plot = TRUE,
  color_var = "seqnames",
  contig_var = NULL
)
}
\arguments{
\item{gr}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{leftstart}{number: left flank start  (relative to range start)}

\item{leftend}{number: left flank  end   (relative to range start)}

\item{rightstart}{number: right flank start (relative to range end)}

\item{rightend}{number: right flank end   (relative to range end)}

\item{bsgenome}{\code{\link[BSgenome]{BSgenome-class}}
Required to update gr$seq if present.}

\item{verbose}{TRUE (default) or FALSE}

\item{plot}{TRUE (default) or FALSE}

\item{color_var}{string (default 'seqnames') var mapped to plot color}

\item{contig_var}{NULL (default) or string: var mapped to plot contig}
}
\value{
\code{\link[GenomicRanges]{GRanges-class}}
}
\description{
Flank left and right and merge overlaps
}
\examples{
bedfile <- system.file('extdata/SRF.bed', package = 'multicrispr')
gr <- bed_to_granges(bedfile, 'mm10', plot = FALSE)
double_flank(gr)
}
