% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data.R
\name{EnsDb.Hsapiens.v99}
\alias{EnsDb.Hsapiens.v99}
\title{Get EnsDb.Hsapiens.v98 from AnnotationHub}
\usage{
EnsDb.Hsapiens.v99()
}
\value{
\code{\link[ensembldb]{EnsDb-class}}
}
\description{
Get EnsDb.Hsapiens.v98 from AnnotationHub
}
\examples{
EnsDb.Hsapiens.v99()
}
