% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data.R
\name{EnsDb.Mmusculus.v98}
\alias{EnsDb.Mmusculus.v98}
\title{Get EnsDb.Mmusculus.v98 from AnnotationHub}
\usage{
EnsDb.Mmusculus.v98()
}
\value{
\code{\link[ensembldb]{EnsDb-class}}
}
\description{
Get EnsDb.Mmusculus.v98 from AnnotationHub
}
\examples{
EnsDb.Mmusculus.v98()
}
