% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_filter_doench_score.R
\name{filter_doench_score}
\alias{filter_doench_score}
\title{Filter on Doench score}
\usage{
filter_doench_score(cas9s, bsgenome, ruleset = 1, filter = 0.3,
  python = NULL, virtualenv = NULL, condaenv = NULL, plot = TRUE,
  verbose = TRUE)
}
\arguments{
\item{cas9s}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{bsgenome}{\code{\link[BSgenome]{BSgenome-class}}}

\item{ruleset}{1 (default) or 2 (requires non-NULL value for either python, virtualenv, or condaenv)}

\item{python}{NULL (default) or python binary path with module azimuth}

\item{virtualenv}{NULL (default) or python virtualenv with module azimuth}

\item{condaenv}{NULL (default) or python condaenv with module azimuth}

\item{plot}{TRUE (default) or FALSE}

\item{verbose}{TRUE (default) or FALSE}

\item{contextseqs}{character vector with 4-23-3 contextseqs}
}
\value{
numeric vector
}
\description{
Score cas9s using ruleset1 (Doench 2014) or ruleset2 (Doench 2016). 
Then filter on set threshold.
}
\details{
ruleset1 is readily available. ruleset2 is accessible after installing 
the python module [azimuth](https://github.com/MicrosoftResearch/Azimuth), 
and specifying a value for either \code{python} (python binary path), 
\code{virtualenv} (python virtual environment dir) or 
\code{condaenv} (python conda environment).
}
\examples{
# Get cas9s
    require(magrittr)
    bedfile <- system.file('extdata/SRF.bed', package = 'multicrispr')
    txdb <- utils::getFromNamespace('TxDb.Mmusculus.UCSC.mm10.knownGene', 
                                    'TxDb.Mmusculus.UCSC.mm10.knownGene')
    bsgenome <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
    targets <- bed_to_granges(bedfile, txdb)  \%>\% 
               double_flank() \%>\%
               add_seq(bsgenome)

# Find cas9s
    cas9s <- find_cas9s(targets)
    
# Compute and filter on Doench score
    cas9s \%>\% filter_doench_score(bsgenome)

}
\references{
Doench 2014, Rational design of highly active sgRNAs for 
CRISPR-Cas9-mediated gene inactivation. Nature Biotechnology,
doi: 10.1038/nbt.3026

Doench 2016, Optimized sgRNA design to maximize activity and minimize 
off-target effects of CRISPR-Cas9. Nature Biotechnology, 
doi: 10.1038/nbt.3437

Python module azimuth: github/MicrosoftResearch/azimuth
}
