% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_filter_no_offtargets.R
\name{filter_no_offtargets}
\alias{filter_no_offtargets}
\title{Filter for no offtargets}
\usage{
filter_no_offtargets(
  cas9s,
  targets,
  bsgenome,
  mismatch = 2,
  offtargetchr = GenomeInfoDb::standardChromosomes(targets),
  plot = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{cas9s}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{targets}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{bsgenome}{\code{\link[BSgenome]{BSgenome-class}}}

\item{mismatch}{number: max number of mismatches to consider}

\item{offtargetchr}{character vector: chromosomes for offtarget analysis}

\item{plot}{logical(1)}

\item{verbose}{logical(1)}
}
\value{
\code{\link[GenomicRanges]{GRanges-class}}
        mcols(GRanges) contains sequences and match counts:
            matches0 = perfect match counts
            matches1 = single mismatch counts
            matches2 = double mismatch counts
}
\description{
Filter for no offtargets
}
\examples{
# Read ranges, extend, and restrict to 'chrY' (to keep example fast)
    require(magrittr)
    bedfile <- system.file('extdata/SRF.bed', package = 'multicrispr')
    targets <- bed_to_granges(bedfile, 'mm10')
    targets \%<>\% extend()
    targets \%<>\% extract(GenomicRanges::seqnames(.)=='chrY')

# Add seqs and find cas9s
    bsgenome <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
    targets \%<>\% add_seq(bsgenome)
    cas9s <- find_crispr_sites(targets)
   
# Filter for offtarget-free cas9s
    filter_no_offtargets(cas9s, targets, bsgenome, 0, offtargetchr = 'chrY')
}
