% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_filter_offtargetfree_sites.R
\name{filter_offtargetfree_sites}
\alias{filter_offtargetfree_sites}
\title{Filter for offtarget-free crispr sites}
\usage{
filter_offtargetfree_sites(
  sites,
  targets,
  bsgenome,
  mismatch = 2,
  include = standardChromosomes(targets),
  plot = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{sites}{\code{\link[GenomicRanges]{GRanges-class}}: crispr sites}

\item{targets}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{bsgenome}{\code{\link[BSgenome]{BSgenome-class}}}

\item{mismatch}{number: max number of mismatches to consider}

\item{include}{character vector: offtarget analysis chromosomes}

\item{plot}{TRUE (default) or FALSE}

\item{verbose}{TRUE (default) or FALSE}
}
\value{
\code{\link[GenomicRanges]{GRanges-class}}
        mcols(GRanges) contains sequences and match counts:
            matches0 = perfect match counts
            matches1 = single mismatch counts
            matches2 = double mismatch counts
}
\description{
Filter for offtarget-free crispr sites
}
\examples{
# Read ranges, extend, and restrict to 'chrY' (to keep example fast)
    require(magrittr)
    bedfile <- system.file('extdata/SRF.bed', package = 'multicrispr')
    targets <- bed_to_granges(bedfile, 'mm10')
    targets \%<>\% extend()
    targets \%<>\% extract(seqnames(.)=='chrY')

# Add seqs and find crispr sites
    bsgenome <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
    targets \%<>\% add_seq(bsgenome)
    sites <- find_crispr_sites(targets, bsgenome)
   
# Filter for offtarget-free sites
    filter_offtargetfree_sites(
        sites, targets, bsgenome, 0, include = 'chrY')
}
