% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_filter_specific.R
\name{add_specificity}
\alias{add_specificity}
\alias{filter_target_specific}
\title{Filter for target-specific spacers}
\usage{
add_specificity(
  spacers,
  targets,
  bsgenome,
  mismatches = 2,
  pam = "NGG",
  outdir = OUTDIR,
  indexedgenomesdir = INDEXEDGENOMESDIR,
  plot = TRUE,
  verbose = TRUE
)

filter_target_specific(
  spacers,
  targets,
  bsgenome = getBSgenome(genome(spacers)[1]),
  mismatches = 2,
  pam = "NGG",
  outdir = OUTDIR,
  indexedgenomesdir = INDEXEDGENOMESDIR,
  plot = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{spacers}{spacer \code{\link[GenomicRanges]{GRanges-class}}}

\item{targets}{target \code{\link[GenomicRanges]{GRanges-class}}}

\item{bsgenome}{\code{\link[BSgenome]{BSgenome-class}}}

\item{mismatches}{number (default 2): max number of mismatches to consider}

\item{pam}{string (default 'NGG'): pam sequence}

\item{outdir}{directory where output is written to}

\item{plot}{TRUE (default) or FALSE}

\item{verbose}{TRUE (default) or FALSE}
}
\description{
Filter for target-specific spacers
}
\examples{
# TFBS example
#-------------
 require(magrittr)
 bsgenome <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
 bedfile  <- system.file('extdata/SRF.bed', package = 'multicrispr')
 targets <- extend(bed_to_granges(bedfile, genome = 'mm10'))
 spacers <- find_spacers(targets, bsgenome)
 # index_genome(bsgenome)
 # spacers \%<>\% add_specificity(targets, bsgenome)
 # spacers \%>\% filter_target_specific(targets, bsgenome)
}
