% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_find_cas9s.R
\name{find_cas9ranges}
\alias{find_cas9ranges}
\title{Find cas9 ranges in targetranges}
\usage{
find_cas9ranges(targetranges, verbose = TRUE)
}
\arguments{
\item{targetranges}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{verbose}{logical(1)}
}
\value{
\code{\link[GenomicRanges]{GRanges-class}}
}
\description{
Find cas9 ranges in targetranges
}
\examples{
bedfile  <- system.file('extdata/SRF.bed', package='multicrispr')
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
targetranges <- slop_fourways(read_bed(bedfile, bsgenome))
find_cas9ranges(targetranges)
}
