% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_find_cas9s.R
\name{find_cas9s}
\alias{find_cas9s}
\title{Find cas9 sites in targetranges}
\usage{
find_cas9s(targetranges, verbose = TRUE)
}
\arguments{
\item{targetranges}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{verbose}{logical(1)}
}
\value{
\code{\link[GenomicRanges]{GRanges-class}}
}
\description{
Find cas9 sites in targetranges
}
\examples{
bedfile  <- system.file('extdata/SRF.bed', package='multicrispr')
targetranges <- slop_fourways(read_bed(bedfile, 'mm10'))
find_cas9s(targetranges)
}
