% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_find_crispr_sites.R
\name{find_crispr_sites}
\alias{find_crispr_sites}
\title{Find crispr sites in targetranges}
\usage{
find_crispr_sites(gr, bsgenome, plot = TRUE, verbose = TRUE)
}
\arguments{
\item{gr}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{bsgenome}{\code{\link[BSgenome]{BSgenome-class}}}

\item{plot}{TRUE (default) or FALSE}

\item{verbose}{TRUE (default) or FALSE}
}
\value{
\code{\link[GenomicRanges]{GRanges-class}}
}
\description{
Find crispr sites in targetranges
}
\examples{
# Read bed into granges and extend
    require(magrittr)
    bsgenome <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
    bedfile  <- system.file('extdata/SRF.bed', package='multicrispr')
    gr <- bed_to_granges(bedfile, 'mm10')
    gr \%<>\% extend(plot = FALSE)
    
# Find crispr sites
    find_crispr_sites(gr, bsgenome)
}
\seealso{
\code{\link{find_prime_sites}} to find prime editing sites
}
