% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_find_cas9s.R
\name{find_flankonly_cas9s}
\alias{find_flankonly_cas9s}
\title{Find flank-only cas9s}
\usage{
find_flankonly_cas9s(bedfile, bsgenome, leftstart = -200, leftend = -1,
  rightstart = 1, rightend = 200, verbose = TRUE)
}
\arguments{
\item{bedfile}{character(1): bed file with target sites}

\item{bsgenome}{BSgenome, e.g. BSgenome.Mmusculus.UCSC.mm10::Mmusculus}

\item{leftstart}{numeric(1): left flank start  (rel. to range start)}

\item{leftend}{numeric(1): left flank end    (rel. to range start)}

\item{rightstart}{numeric(1): right flank start (rel. to range end)}

\item{rightend}{numeric(1): right flank end   (rel. to range end)}

\item{verbose}{logical(1)}

\item{bsgenome}{BSgenome object}
}
\description{
Find flank-only cas9s
}
\examples{
bedfile <- system.file('extdata/SRF_sites.bed', package = 'crisprapex')
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
find_flankonly_cas9s(bedfile, bsgenome)
}
