% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_pe3.R
\name{find_nicking_spacers}
\alias{find_nicking_spacers}
\title{Find nicking spacers}
\usage{
find_nicking_spacers(
  pespacers,
  bsgenome,
  plot = TRUE,
  indexedgenomesdir = INDEXEDGENOMESDIR
)
}
\arguments{
\item{pespacers}{GRanges: prime editing spacers}

\item{bsgenome}{BSgenome}

\item{plot}{TRUE (default) or FALSE}

\item{indexedgenomesdir}{string: directory with indexed genomes}
}
\value{
GRanges: nicking spacers
}
\description{
Find nicking spacers
}
\examples{
bsgenome <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38  
gr <- char_to_granges(c(
        PRNP = 'chr20:4699600:+',             # snp: prion disease
        HBB  = 'chr11:5227002:-',             # snp: sickle cell anemia
        HEXA = 'chr15:72346580-72346583:-',   # del: tay sachs disease
        CFTR = 'chr7:117559593-117559595:+'), # ins: cystic fibrosis
        bsgenome)
pespacers <- find_pe_spacers(gr, bsgenome)
# index_genome(bsgenome)
# pespacers \%<>\% add_offtargets(bsgenome, mismatches=0)
find_nicking_spacers(pespacers, bsgenome)
}
