% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_find_cas9s.R
\name{find_offtargetfree_cas9s}
\alias{find_offtargetfree_cas9s}
\title{Find offtarget-free cas9 sites in targetranges}
\usage{
find_offtargetfree_cas9s(targetranges, mismatch = 2,
  offtargetchr = canonicalseqlevels(targetranges), verbose = TRUE)
}
\arguments{
\item{targetranges}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{mismatch}{number: max number of mismatches to consider}

\item{offtargetchr}{character vector: chromosomes for offtarget analysis, 
probably generated with genomeInfoDb::seqlevels(targetranges) or 
canonicalseqlevels(targetranges)}

\item{verbose}{logical(1)}
}
\value{
\code{\link[GenomicRanges]{GRanges-class}}
        mcols(GRanges) contains sequences and match counts:
            matches0 = perfect match counts
            matches1 = single mismatch counts
            matches2 = double mismatch counts
}
\description{
Find offtarget-free cas9 sites in targetranges
}
\examples{
# Note: restricting example to 'chrY' only to keep it fast
bedfile  <- system.file('extdata/SRF.bed', package = 'multicrispr')
targetranges <- read_bed(bedfile, 'mm10', plot = FALSE) \%>\% 
                extract(GenomeInfoDb::seqnames(.)=='chrY') \%>\% 
                extend()
find_offtargetfree_cas9s(targetranges, 0, offtargetchr = 'chrY')
}
