% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_find_pe_sites.R
\name{find_pe_sites}
\alias{find_pe_sites}
\title{Find prime editing sites}
\usage{
find_pe_sites(
  gr,
  bsgenome,
  fix = BSgenome::getSeq(bsgenome, gr, as.character = TRUE),
  nprimer = 13,
  nrt = 16,
  plot = TRUE
)
}
\arguments{
\item{gr}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{bsgenome}{\code{\link[BSgenome]{BSgenome-class}}}

\item{fix}{character vector: fixing sequence on '+' strand}

\item{nprimer}{n primer nucleotides (default 13, max 17)}

\item{nrt}{n RT nucleotides (default 16, recommended 10-16)}

\item{plot}{TRUE (default) or FALSE}
}
\description{
Find prime editing sites
}
\examples{
bsgenome <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38  
gr <- GenomicRanges::GRanges(
        seqnames = c(PRNP = 'chr20:4699600',             # snp
                     HBB  = 'chr11:5227002',             # snp
                     HEXA = 'chr15:72346580-72346583',   # del
                     CFTR = 'chr7:117559593-117559595'), # ins
        strand   = c(PRNP = '+', HBB = '-', HEXA = '-', CFTR = '+'), 
        seqinfo  = BSgenome::seqinfo(bsgenome))
BSgenome::getSeq(bsgenome, gr)
}
