% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_find_cas9s.R
\name{find_specific_cas9s}
\alias{find_specific_cas9s}
\title{Find cas9sites specfic for targetranges}
\usage{
find_specific_cas9s(targetranges, verbose = TRUE)
}
\arguments{
\item{targetranges}{GenomicRanges::GRanges}

\item{verbose}{logical(1)}
}
\value{
GenomicRanges::GRanges
}
\description{
Find cas9sites specfic for targetranges
}
\examples{
require(magrittr)
bedfile <- system.file('extdata/SRF_sites.bed', package = 'crisprapex')
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
targetranges <- read_bed(bedfile, bsgenome) \%>\% flank_fourways()
}
