% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_find_cas9s.R
\name{find_targetonly_cas9s}
\alias{find_targetonly_cas9s}
\title{Find cas9 sites in given granges}
\usage{
find_targetonly_cas9s(targetranges, tabooranges, verbose = TRUE)
}
\arguments{
\item{targetranges}{GenomicRanges::GRanges}

\item{tabooranges}{GenomicRanges::GRanges}

\item{verbose}{logical(1)}
}
\value{
GenomicRanges::GRanges
}
\description{
Find cas9 sites in given granges
}
\examples{
require(magrittr)
bedfile  <- system.file('extdata/SRF_sites.bed', package='crisprapex')
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
targetranges <- read_bed(bedfile, bsgenome) \%>\% flank_fourways()
tabooranges  <- read_bed(bedfile, bsgenome) \%>\% slop_fourways()
find_targetonly_cas9s(targetranges, tabooranges)
}
