% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_flank_slop.R
\name{flank_fourways}
\alias{flank_fourways}
\title{Flank fourways}
\usage{
flank_fourways(ranges, bsgenome, leftstart = -200, leftend = -1,
  rightstart = 1, rightend = 200, verbose = TRUE)
}
\arguments{
\item{ranges}{data.table(chr, start, end, strand)}

\item{bsgenome}{BSgenome, e.g. BSgenome.Mmusculus.UCSC.mm10::Mmusculus}

\item{leftstart}{numeric(1): left flank start  (from range start)}

\item{leftend}{numeric(1): left flank  end   (from range start)}

\item{rightstart}{numeric(1): right flank start (from range end)}

\item{rightend}{numeric(1): right flank end   (from range end)}

\item{verbose}{logical(1): report?}
}
\value{
data.table(chr, start, end, strand)
}
\description{
Flank left and right, for both strands, and merge overlaps
}
\examples{
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
bedfile <- system.file('extdata/SRF_sites.bed', package = 'crisprapex')
ranges <- read_bed(bedfile)
flank_fourways(ranges, bsgenome)
}
