% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_xxx_to_granges.R
\name{genes_to_granges}
\alias{genes_to_granges}
\alias{genefile_to_granges}
\title{Convert geneids into GRanges}
\usage{
genes_to_granges(geneids, txdb, bsgenome = NULL, plot = TRUE)

genefile_to_granges(file, txdb, bsgenome = NULL, plot = TRUE)
}
\arguments{
\item{geneids}{Entrez Gene identifier vector}

\item{txdb}{\code{\link[GenomicFeatures]{TxDb-class}} or 
\code{\link[ensembldb]{EnsDb-class}}}

\item{plot}{TRUE or FALSE}

\item{file}{Entrez Gene identifier file (one per row)}

\item{BSgenome}{\code{\link[BSgenome]{BSgenome-class}} or NULL. 
If specified, used to extract and add sequence to GRanges.}

\item{addseq}{logical(1)}
}
\value{
\code{\link[GenomicRanges]{GRanges-class}}
}
\description{
Convert geneids into GRanges
}
\examples{

# Entrez
#-------
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
genefile <- system.file('extdata/SRF.entrez', package='multicrispr')
geneids  <- as.character(read.table(genefile)[[1]])
txdb       <- utils::getFromNamespace('TxDb.Mmusculus.UCSC.mm10.knownGene',
                                    'TxDb.Mmusculus.UCSC.mm10.knownGene')
gr <- genes_to_granges(geneids, txdb)
gr <- genes_to_granges(geneids, txdb, bsgenome)
gr <- genefile_to_granges(genefile, txdb)
gr <- genefile_to_granges(genefile, txdb, bsgenome)

# Ensembl
#--------
txdb <- EnsDb.Mmusculus.v98()
genefile <- system.file('extdata/SRF.ensembl', package='multicrispr')
geneids <- as.character(read.table(genefile)[[1]])
gr <- genes_to_granges(geneids, txdb)
gr <- genes_to_granges(geneids, txdb, bsgenome)
gr <- genefile_to_granges(genefile, txdb)
gr <- genefile_to_granges(genefile, txdb, bsgenome)
}
\seealso{
\code{\link{annotate_granges}} (inverse operation)
}
