% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_read_assert.R
\name{get_bsgenome}
\alias{get_bsgenome}
\title{Get BSgenome}
\usage{
get_bsgenome(gr)
}
\arguments{
\item{gr}{\code{\link[GenomicRanges]{GRanges-class}}}
}
\value{
BSgenome
}
\description{
Get BSgenome
}
\examples{
bedfile  <- system.file('extdata/SRF.bed', package='multicrispr')
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
gr <- read_bed(bedfile, bsgenome)
get_bsgenome(gr)
}
