% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_filter_offtargetfree_sites.R
\name{include_to_exclude}
\alias{include_to_exclude}
\title{Convert "included" into "excluded" seqnames}
\usage{
include_to_exclude(
  bsgenome,
  include = GenomeInfoDb::standardChromosomes(bsgenome)
)
}
\arguments{
\item{bsgenome}{\code{\link[BSgenome]{BSgenome-class}}}

\item{include}{character vector with names of included chromosomes}
}
\value{
character vector
}
\description{
Convert "included" into "excluded" seqnames
}
\details{
Convert "included" into "excluded" seqnames. 
         Do this in regex format as required by 
\code{\link[BSgenome]{BSgenome-utils}}) \code{vcountPDict} and 
\code{vmatchPDict}
}
\examples{
bsgenome <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38
include_to_exclude(bsgenome)
}
