% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_filter_specific.R
\name{index_genome}
\alias{index_genome}
\title{Index genome}
\usage{
index_genome(
  bsgenome,
  indexedgenomesdir = INDEXEDGENOMESDIR,
  overwrite = FALSE
)
}
\arguments{
\item{bsgenome}{\code{\link[BSgenome]{BSgenome-class}}}

\item{indexedgenomesdir}{string: directory with bowtie-indexed genome}

\item{overwrite}{logical(1)}
}
\value{
invisible(genomdir)
}
\description{
Bowtie index genome
}
\examples{
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
#index_genome(bsgenome)
bsgenome <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38
#index_genome(bsgenome)
}
