% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_alignment_offtargetanalysis.R
\name{index_targets}
\alias{index_targets}
\title{Index targets}
\usage{
index_targets(targets, bsgenome, outdir = "~/.multicrispr/bowtie/target")
}
\arguments{
\item{targets}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{bsgenome}{\code{\link[BSgenome]{BSgenome-class}}}

\item{outdir}{string: output directory}
}
\value{

}
\description{
Index targets
}
\examples{
require(magrittr)
bedfile  <- system.file('extdata/SRF.bed', package = 'multicrispr')
targets0 <- bed_to_granges(bedfile, genome = 'mm10')
targets  <- double_flank(targets0)
index_targets(targets, bsgenome, outdir = outdir)
}
