% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_utilities.R
\name{make_unique_names}
\alias{make_unique_names}
\title{Make unique names}
\usage{
make_unique_names(x, prefix = "T")
}
\arguments{
\item{x}{vector}
}
\value{
character vector with unique names
}
\description{
Make unique names
}
