% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_add_genome_matches.R
\name{match_seqs}
\alias{match_seqs}
\alias{match_crisprs}
\alias{add_genome_matches}
\alias{add_target_matches}
\title{Add match counts}
\usage{
match_seqs(
  seqs,
  indexdir,
  norc,
  outfile,
  count_vars = c("MM0", "MM1", "MM2"),
  verbose = TRUE
)

match_crisprs(
  spacers,
  indexdir,
  norc,
  outfile,
  pam = "NGG",
  count_vars = c("MM0", "MM1", "MM2"),
  verbose = TRUE
)

add_genome_matches(spacers, indexdir, outdir, pam = "NGG", verbose = TRUE)

add_target_matches(spacers, indexdir, outdir, pam = "NGG", verbose = TRUE)
}
\arguments{
\item{seqs}{character vector: sequences to match against indexed ref}

\item{indexdir}{string: dir containing indexed reference.
This can be an indexed genome( \code{\link{index_genome}}
It can also be indexed targets (\code{\link{index_targets}})}

\item{norc}{TRUE or FALSE: whether to run bowtie also with revcompls
Generally TRUE for genome and FALSE for target matches, 
because target ranges generally include both strands.}

\item{outfile}{string: file where to output bowtie results}

\item{count_vars}{character(3): how to name count variables}

\item{verbose}{TRUE (default) or FALSE}

\item{spacers}{\code{\link[GenomicRanges]{GRanges-class}} with crispr 
sites to be matched against indexed ref}

\item{pam}{string (default 'NGG') pam pattern to expand}

\item{outdir}{string: dir where to output bowtie results}
}
\description{
Count matches to indexed target/genome and add to GRanges
}
\details{
\code{match_seqs} matches sequences against (indexed) target and genome
\code{add_crispr_matches} expands iupac amgiguities in the pam sequence, 
matches all resulting sequences against (indexes) target and genome, 
adds match counts to GRanges object, and then returns it
}
\examples{
# PE example
#-----------
 require(magrittr)
 bsgenome <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38  
 genomedir <- '~/.multicrispr/bowtie/genome/BSgenome.Hsapiens.UCSC.hg38'
 # index_genome(bsgenome, genomedir) # one time effort - takes few h
 gr <- char_to_granges(c(PRNP = 'chr20:4699600:+',             # snp
                         HBB  = 'chr11:5227002:-',             # snp
                         HEXA = 'chr15:72346580-72346583:-',   # del
                         CFTR = 'chr7:117559593-117559595:+'), # ins
                       bsgenome)
 spacers <- find_pe_spacers(gr, bsgenome)
 outdir <- '~/.multicrispr/bowtie'
 outfile <- file.path(outdir, 'crispr_to_ref.txt')
 seqs <- paste0(spacers$crisprspacer, spacers$crisprpam)
 match_seqs(seqs,       genomedir, norc=FALSE, outfile=outfile)
 match_crisprs(spacers, genomedir, norc=FALSE, outfile=outfile)
 add_genome_matches(spacers, genomedir, outdir)
    
# TFBS example
#-------------
 bsgenome <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
 gdir <- '~/.multicrispr/bowtie/genome/BSgenome.Mmusculus.UCSC.mm10'
 # index_genome(bsgenome, gdir) # one time effort - takes few h
 bedfile  <- system.file('extdata/SRF.bed', package = 'multicrispr')
 targets0 <- bed_to_granges(bedfile, genome = 'mm10')
 targets  <- double_flank(targets0)
 
 tdir <- index_targets(targets, bsgenome)
 spacers <- find_spacers(targets, bsgenome)
 spacers \%<>\% extract(1:10) # only to keep example fast, works also else
 outdir <- '~/.multicrispr/bowtie'
 outfile <- file.path(outdir, 'crispr_to_ref.txt')
    
 crisprseqs <- unique(paste0(spacers$crisprspacer, spacers$crisprpam))
 match_seqs(crisprseqs, gdir, norc=FALSE, outfile=outfile)
 match_crisprs(spacers, gdir, norc = FALSE, outfile = outfile)
 spacers \%>\% add_target_matches(tdir, outdir)
 spacers \%>\% add_genome_matches(gdir, outdir)
}
\seealso{
\code{\link{index_genome}}, \code{\link{index_targets}}
}
