% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_flank_slop.R
\name{merge_overlaps}
\alias{merge_overlaps}
\title{Merge overlaps}
\usage{
merge_overlaps(x, bsgenome, verbose = TRUE)
}
\arguments{
\item{x}{data.table(chr, start, end, strand)}

\item{bsgenome}{BSgenome, e.g. BSgenome.Mmusculus.UCSC.mm10::Mmusculus}

\item{verbose}{logical(1)}
}
\value{
data.table(chr, start, end)
}
\description{
Merge overlaps
}
\examples{
require(magrittr)
(rangesdt <- data.table::data.table(
            chr    = rep('chr1', 4), 
            start  = c(1,    5,   5,  20), 
            end    = c(10,  15,  15,  30), 
            strand = c('+', '+', '-', '-')))
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
rangesdt \%>\% merge_overlaps(bsgenome)
}
