% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_plot.R
\name{plot_karyogram}
\alias{plot_karyogram}
\alias{plot_tracks}
\alias{plot_intervals}
\title{Plot GRanges(List)}
\usage{
plot_karyogram(grangeslist, title = unique(genome(grangeslist)))

plot_tracks(grangeslist)

plot_intervals(grangeslist)
}
\arguments{
\item{grangeslist}{\code{\link[GenomicRanges]{GRanges-class}} or 
\code{\link[GenomicRanges]{GRangesList-class}}}

\item{title}{plot title}
}
\value{
list (plot_karyogram) or ggplot (plot_intervals)
}
\description{
Plot GRanges(List)
}
\details{
Visualize a GRanges(List) as karyogram (plot_karyogram), 
genome tracks (plot_tracks) or intervals (plot_intervals)
}
\examples{
# Plot GRanges
  bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
  bedfile <-  system.file('extdata/SRF.bed',  package = 'multicrispr')
  granges   <- read_bed(bedfile,  bsgenome, plot = FALSE)
  plot_karyogram(granges)
  plot_intervals(granges)
  
# Plot GRangesList
  flanks  <- left_flank(granges)
  grangeslist <- GenomicRanges::GRangesList(sites = granges, flanks = flanks)
  plot_karyogram(grangeslist)
  plot_intervals(grangeslist)
}
