% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_find_cas9s.R
\name{range2seq}
\alias{range2seq}
\title{Convert range to sequence}
\usage{
range2seq(chr, start, end, strand, bsgenome)
}
\arguments{
\item{chr}{character vector: c('chr1', 'chr1', ...)}

\item{start}{numeric vector: start positions}

\item{end}{numeric vector: end positions}

\item{strand}{character vector: +/- values}

\item{bsgenome}{BSgenome object}
}
\value{
character vector: sequences
}
\description{
Convert range to sequence
}
\examples{
range2seq(
    chr      = c('chr2', 'chr1', 'chr17'),
    start    = c(83972889, 85603908, 70963622),
    end      = c(83972905, 85603924, 70963638),
    strand   = c('+', '-', '-'),
    bsgenome = BSgenome.Mmusculus.UCSC.mm10::Mmusculus)
}
