% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_functions.R
\name{rangedt2seqdt}
\alias{rangedt2seqdt}
\title{Convert range data.table into sequence data.table}
\usage{
rangedt2seqdt(rangedt, bsgenome, verbose = TRUE)
}
\arguments{
\item{rangedt}{data.table(chr, start, end)}

\item{verbose}{logical(1)}
}
\value{
data.table
}
\description{
Convert range data.table into sequence data.table
}
\examples{
rangedt <- data.table::data.table(
    chr      = c('chr2', 'chr1', 'chr17'),
    start    = c(83972889, 85603908, 70963622),
    end      = c(83972905, 85603924, 70963638),
    strand   = c('+', '-', '-'))
    
rangedt2seqdt(rangedt, BSgenome.Mmusculus.UCSC.mm10::Mmusculus) 
}
