% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range2seq.R
\name{read_bed}
\alias{read_bed}
\title{Read bedfile as data.table}
\usage{
read_bed(bedfile, verbose = TRUE)
}
\arguments{
\item{bedfile}{file path}

\item{verbose}{logical(1)}
}
\description{
Read bedfile as data.table
}
\examples{
bedfile <- system.file('extdata/SRF_sites.bed', package = 'cas9tbflanks')
read_bed(bedfile)
}
