% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_read_flank_slop.R
\docType{methods}
\name{reduce,data.table-method}
\alias{reduce,data.table-method}
\title{Range merge overlaps}
\usage{
\S4method{reduce}{data.table}(x, verbose = TRUE)
}
\arguments{
\item{x}{data.table(chr, start, end, strand)}

\item{verbose}{logical(1)}
}
\value{
data.table(chr, start, end)
}
\description{
Range merge overlaps
}
\examples{
(ranges <- data.table::data.table(
            chr    = rep('chr1', 4), 
            start  = c(1,    5,   5,  20), 
            end    = c(10,  15,  15,  30), 
            strand = c('+', '+', '-', '-')))
reduce(ranges)
}
