% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_flank_slop.R
\name{right_flank}
\alias{right_flank}
\title{Right flank}
\usage{
right_flank(ranges, bsgenome, rightstart = 1, rightend = 200,
  verbose = TRUE)
}
\arguments{
\item{ranges}{data.table(chr, start, end)}

\item{bsgenome}{BSgenome, e.g. BSgenome.Mmusculus.UCSC.mm10::Mmusculus}

\item{rightstart}{flank start relative to range start}

\item{verbose}{logical(1)}

\item{leftend}{flank end   relative to range start}
}
\value{
data.table
}
\description{
Right flank
}
\examples{
require(magrittr)
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
bedfile <- system.file('extdata/SRF_sites.bed', package = 'crisprapex')
ranges <- read_bed(bedfile)
ranges \%>\% head(1)
ranges \%>\% head(1) \%>\% right_flank(bsgenome)
}
