% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_find_cas9s.R
\name{rm_center_cas9s}
\alias{rm_center_cas9s}
\title{Rm center cas9s}
\usage{
rm_center_cas9s(flank_cas9s, center_cas9s, verbose = TRUE)
}
\arguments{
\item{flank_cas9s}{data.table(cas9seq, ...)}

\item{center_cas9s}{data.table(cas9seq, ...)}

\item{verbose}{logical(1)}
}
\value{
subset of flank_cas9s
}
\description{
Rm center cas9s
}
\examples{
require(magrittr)
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
bedfile <- system.file('extdata/SRF_sites.bed', package='crisprapex')
tbranges <- read_bed(bedfile)
flank9s  <- tbranges \%>\% flank_fourways(bsgenome) \%>\% find_cas9s(bsgenome)
center9s <- tbranges \%>\% slop_fourways(bsgenome)  \%>\% find_cas9s(bsgenome)
flank9s \%>\% rm_center_cas9s(center9s)

}
