% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_scan_genome.R
\name{scan_genome}
\alias{scan_genome}
\title{Scan genome}
\usage{
scan_genome(cas9seqs, bsgenome, mismatch, verbose = TRUE)
}
\arguments{
\item{cas9seqs}{character(1)}

\item{bsgenome}{BSgenome object}
}
\value{
data.table
}
\description{
Scan genome
}
\examples{
require(magrittr)
bsgenome <-  BSgenome.Mmusculus.UCSC.mm10::Mmusculus
bedfile <- system.file('extdata/SRF_sites.bed', package = 'crisprapex')
cas9seqs <- read_bed(bedfile)        \%>\%
            head(30)                 \%>\%  
            slop_fourways(bsgenome)  \%>\% 
            find_cas9s(bsgenome)     \%>\% 
            extract2('cas9seq')
}
