% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_score_cas9s.R
\name{score_cas9ranges}
\alias{score_cas9ranges}
\title{Calc cas9 ontargetscores}
\usage{
score_cas9ranges(chr, start, end, strand, bsgenome, verbose = TRUE)
}
\arguments{
\item{bsgenome}{BSGenome object, e.g. BSgenome.Mmusculus.UCSC.mm10::Mmusculus}

\item{cas9dt}{data.table(chr, cas9start, cas9end, strand)}
}
\value{
data.table(chr, cas9start, cas9end, strand)
}
\description{
Calc cas9 ontargetscores
}
\examples{
if (reticulate::py_module_available('azimuth')){
    require(magrittr)
    bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
    bedfile <- system.file('extdata/SRF_sites.bed', package = 'crisprapex')
    cas9dt <-   read_bed(bedfile)    \%>\% 
                head(3)              \%>\% 
                slop_fourways()      \%>\% 
                find_cas9s(bsgenome)
    cas9dt [ , score_cas9ranges(chr, cas9start, cas9end, strand, bsgenome) ]
    cas9dt \%>\% add_ontargetscores(chr, cas9start, cas9end, strand, bsgenome)
}
}
