% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_score_cas9s.R
\name{score_contextseqs}
\alias{score_contextseqs}
\title{Calculate azimuth ontargetscores}
\usage{
score_contextseqs(contextseqs, verbose = TRUE)
}
\arguments{
\item{contextseqs}{character vector}

\item{verbose}{logical(1)}
}
\value{
numeric vector
}
\description{
Calculate azimuth ontargetscores
}
\examples{
if (reticulate::py_module_available('azimuth')){
    contextseqs <- c('TGCCCTTATATTGTCTCCAGCAGAAGGTGT',
                     'TGCCCTTATATTGTCTCCAGCAGAAGGTGT',
                     'CCAAATATTGTCAAGTTGACAACCAGGAAT')
    calc_ontargetscore(contextseqs)
}
}
