% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_score_crispr_sites.R
\name{score_crispr_sites}
\alias{score_crispr_sites}
\title{Score crispr sites}
\usage{
score_crispr_sites(
  sites,
  bsgenome,
  method = c("Doench2014", "Doench2016")[1],
  python = NULL,
  virtualenv = NULL,
  condaenv = NULL,
  verbose = TRUE
)
}
\arguments{
\item{sites}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{bsgenome}{\code{\link[BSgenome]{BSgenome-class}}}

\item{method}{'Doench2014' (default) or 'Doench2016'
(requires non-NULL argument python, virtualenv, or condaenv)}

\item{python}{NULL (default) or python binary path with module azimuth}

\item{virtualenv}{NULL (default) or python virtualenv with module azimuth}

\item{condaenv}{NULL (default) or python condaenv with module azimuth}

\item{verbose}{TRUE (default) or FALSE}
}
\value{
numeric vector
}
\description{
Score crispr sites with Doench2014 or Doench2016 model.
}
\details{
Doench2014 is readily available. 
Doench2016 is available after installing python module 
[azimuth](https://github.com/MicrosoftResearch/Azimuth) (see examples).
}
\examples{
# Define targets
    require(magrittr)
    bedfile  <- system.file('extdata/SRF.bed', package = 'multicrispr')
    bsgenome <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
    targets <- bed_to_granges(bedfile, 'mm10')  \%>\% 
               double_flank() \%>\%
               add_seq(bsgenome)

# Find sites
    sites <- find_crispr_sites(targets)
    
# Score with Doench2014
    score_crispr_sites(sites[1:10], bsgenome)
        
# Score with Doench2016
    # First install python module azimuth, perhaps in a conda env:
        # install conda for python 2.7
        # conda create --name azimuthenv python=2.7
        # conda activate azimuthenv
        # pip install azimuth
        # pip install scikit-learn==0.17.1
        
    # Then call score_crispr_sites with reference to conda env
        # score_crispr_sites(sites[1:10], bsgenome, 'Doench2016', 
        #             condaenv = 'azimuthenv')

}
\references{
Doench 2014, Rational design of highly active sgRNAs for 
CRISPR-Cas9-mediated gene inactivation. Nature Biotechnology,
doi: 10.1038/nbt.3026

Doench 2016, Optimized sgRNA design to maximize activity and minimize 
off-target effects of CRISPR-Cas9. Nature Biotechnology, 
doi: 10.1038/nbt.3437

Python module azimuth: github/MicrosoftResearch/azimuth
}
