% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_score_spacers.R
\name{score_spacers}
\alias{score_spacers}
\title{Score spacers}
\usage{
score_spacers(
  spacers,
  bsgenome,
  method = c("Doench2014", "Doench2016")[1],
  python = NULL,
  virtualenv = NULL,
  condaenv = NULL,
  verbose = TRUE
)
}
\arguments{
\item{spacers}{\code{\link[GenomicRanges]{GRanges-class}}: spacers}

\item{bsgenome}{\code{\link[BSgenome]{BSgenome-class}}}

\item{method}{'Doench2014' (default) or 'Doench2016'
(requires non-NULL argument python, virtualenv, or condaenv)}

\item{python}{NULL (default) or python binary path with module azimuth}

\item{virtualenv}{NULL (default) or python virtualenv with module azimuth}

\item{condaenv}{NULL (default) or python condaenv with module azimuth}

\item{verbose}{TRUE (default) or FALSE}
}
\value{
numeric vector
}
\description{
Score spacers with Doench2014 or Doench2016
}
\details{
Doench2014 is readily available. 
Doench2016 is available after installing python module 
[azimuth](https://github.com/MicrosoftResearch/Azimuth) (see examples).
}
\examples{
# PE example
#-----------
    require(magrittr)
    bsgenome <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38  
    gr <- GenomicRanges::GRanges(
              seqnames = c(PRNP = 'chr20:4699600',             # snp
                           HBB  = 'chr11:5227002',             # snp
                           HEXA = 'chr15:72346580-72346583',   # del
                           CFTR = 'chr7:117559593-117559595'), # ins
              strand   = c(PRNP = '+', HBB = '-', HEXA = '-', CFTR = '+'), 
              seqinfo  = BSgenome::seqinfo(bsgenome))
    spacers <- find_spacers(extend_for_pe(gr), bsgenome)
    score_spacers(spacers, bsgenome, 'Doench2014')
        # conda create --name azimuthenv python=2.7
        # conda activate azimuthenv
        # pip install azimuth
        # pip install scikit-learn==0.17.1
    # score_spacers(spacers, bsgenome, 'Doench2016', condaenv = 'azimuthenv')

# TFBS example
#-------------
    bedfile  <- system.file('extdata/SRF.bed', package = 'multicrispr')
    bsgenome <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
    gr <- extend(bed_to_granges(bedfile, 'mm10'))
    spacers <- find_spacers(gr, bsgenome)
    score_spacers(spacers, bsgenome, 'Doench2014')
    # score_spacers(spacers, bsgenome, 'Doench2016')

}
\references{
Doench 2014, Rational design of highly active sgRNAs for 
CRISPR-Cas9-mediated gene inactivation. Nature Biotechnology,
doi: 10.1038/nbt.3026

Doench 2016, Optimized sgRNA design to maximize activity and minimize 
off-target effects of CRISPR-Cas9. Nature Biotechnology, 
doi: 10.1038/nbt.3437

Python module azimuth: github/MicrosoftResearch/azimuth
}
