% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_xxx_to_granges.R
\name{seqs}
\alias{seqs}
\title{Convert GRanges into Sequences}
\usage{
seqs(gr, bsgenome)
}
\arguments{
\item{gr}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{bsgenome}{\code{\link[BSgenome]{BSgenome-class}}}
}
\value{
character vector
}
\description{
Convert GRanges into Sequences
}
\examples{
bedfile  <- system.file('extdata/SRF.bed', package = 'multicrispr')
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
gr <- bed_to_granges(bedfile, bsgenome)
seqs(gr[1:3], bsgenome)
}
