% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_find_cas9s.R
\docType{methods}
\name{sequence}
\alias{sequence}
\alias{sequence,GRanges-method}
\alias{sequence<-}
\alias{sequence<-,GRanges,character-method}
\alias{sequence<-,GRanges,DNAStringSet-method}
\title{Get/set sequence values}
\usage{
sequence(object)

\S4method{sequence}{GRanges}(object)

sequence(object) <- value

\S4method{sequence}{GRanges,character}(object) <- value

\S4method{sequence}{GRanges,DNAStringSet}(object) <- value
}
\arguments{
\item{object}{GenomicRanges::GRanges}

\item{value}{character vector or DNAStringSet}
}
\value{
DNAStringSet (get) or GRanges (set)
}
\description{
Get/set sequence values
}
\examples{
bedfile  <- system.file('extdata/SRF_sites.bed', package = 'crisprapex')
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
granges <- read_bed(bedfile, bsgenome)

sequence(granges)                       # sequence retrieved from bsgenome
sequence(granges) <- sequence(granges)  # sequence stored in granges
sequence(granges)                       # sequence retrieved from granges
}
