% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_flank_slop.R
\name{slop}
\alias{slop}
\title{Slop (i.e. extend left/right)}
\usage{
slop(granges, leftstart = -22, rightend = 22, verbose = TRUE)
}
\arguments{
\item{granges}{GenomicRanges::GRanges}

\item{leftstart}{flank start relative to range start}

\item{rightend}{flank end   relative to range start}

\item{verbose}{logical(1)}
}
\value{
GenomicRanges::GRanges
}
\description{
Slop (i.e. extend left/right)
}
\examples{
require(magrittr)
bedfile <- system.file('extdata/SRF_sites.bed', package = 'crisprapex')
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
granges <- read_bed(bedfile, bsgenome)
granges \%>\% head(1)
granges \%>\% head(1) \%>\% slop()
}
