% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_flank_slop.R
\name{slop_fourways}
\alias{slop_fourways}
\title{Slop granges for both strands, merging overlaps}
\usage{
slop_fourways(gr, leftstart = -22, rightend = 22, plot = TRUE,
  verbose = TRUE)
}
\arguments{
\item{gr}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{leftstart}{number}

\item{rightend}{number}

\item{plot}{logical(1)}

\item{verbose}{logical(1)}
}
\value{
\code{\link[GenomicRanges]{GRanges-class}}
}
\description{
Slop granges for both strands, merging overlaps
}
\examples{
bedfile <- system.file('extdata/SRF.bed', package = 'multicrispr')
gr <- read_bed(bedfile, 'mm10', plot = FALSE)
slop_fourways(gr)
}
