% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_manipulate_ranges.R
\name{straddle}
\alias{straddle}
\title{Straddle GRanges}
\usage{
straddle(
  gr,
  leftstart = NULL,
  leftend = NULL,
  rightstart = NULL,
  rightend = NULL,
  bsgenome = NULL,
  verbose = TRUE,
  plot = TRUE,
  ...
)
}
\arguments{
\item{gr}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{leftstart}{number: left start (relative to range start)}

\item{leftend}{number: left end   (relative to range start)}

\item{rightstart}{number: right start (relative to range end)}

\item{rightend}{number: right end   (relative to range end)}

\item{bsgenome}{\code{\link[BSgenome]{BSgenome-class}}
Required to update gr$seq if present.}

\item{verbose}{TRUE (default) or FALSE}

\item{plot}{TRUE (default) or FALSE}

\item{...}{\code{\link{plot_intervals}} arguments}
}
\value{
a \code{\link[GenomicRanges]{GRanges-class}}
}
\description{
Straddle GRanges
}
\examples{
# HBB snp: sickle cell variant (T -> A)
    require(magrittr)
    bsgenome <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38
    bsinfo <- BSgenome::seqinfo(bsgenome)
    gr <- GenomicRanges::GRanges(
             'chr11:5227002-5227002', strand = '-', seqinfo = bsinfo)
    (gr \%<>\% add_seq(bsgenome))
    gr \%>\% straddle(leftstart = -22, rightend = 22, bsgenome = bsgenome)

# SRF binding sites
    bedfile <- system.file('extdata/SRF.bed', package = 'multicrispr')
    gr <- bed_to_granges(bedfile, 'mm10', plot = FALSE)
    straddle(gr, leftstart=-200, leftend=-1)    # left flank
    straddle(gr, rightstart=1,   rightend=200)  # right flank
    straddle(gr, leftstart=-200, rightend=200)  # extend
}
\author{
Aditya Bhagwat, after discussing with Michael Lawrence on bioc-devel
}
