% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_manipulate_ranges.R
\name{up_flank}
\alias{up_flank}
\alias{down_flank}
\alias{extend}
\title{Extend or Flank GRanges}
\usage{
up_flank(
  gr,
  start = -200,
  end = -1,
  stranded = TRUE,
  bsgenome = NULL,
  verbose = FALSE,
  plot = FALSE,
  ...
)

down_flank(
  gr,
  start = 1,
  end = 200,
  stranded = TRUE,
  bsgenome = NULL,
  verbose = FALSE,
  plot = FALSE,
  ...
)

extend(
  gr,
  start = -22,
  end = 22,
  stranded = TRUE,
  bsgenome = NULL,
  verbose = FALSE,
  plot = FALSE,
  ...
)
}
\arguments{
\item{gr}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{start}{(pos or neg) number: relative start position (see details)}

\item{end}{(pos or neg) number: relative end position   (see details)}

\item{stranded}{TRUE (default) or FALSE: consider strand information}

\item{bsgenome}{NULL (default) or \code{\link[BSgenome]{BSgenome-class}}.
Required to update gr$seq if present.}

\item{verbose}{TRUE or FALSE (default)}

\item{plot}{TRUE or FALSE (default)}

\item{...}{passed to \code{\link{plot_intervals}}}
}
\value{
a \code{\link[GenomicRanges]{GRanges-class}}
}
\description{
Returns extensions, upstream flanks, or downstream flanks
}
\details{
Operations are performed in a strand-aware fashion by default, but can be 
made strand-agnostic by setting stranded=FALSE.

Return values (note: if stranded=FALSE, only + versions are returned)
    up_flank  grstart + start  -->  grstart + end   (+)
              grend - end     <--   grend - start   (-)
           
  down_flank  grend + start    -->  grend + end     (+) 
              grstart - end    <--  grstart - start (-)
              
      extend  grstart + start  -->  grend + end     (+)
              grstart - end    <--  grend - start   (-)
}
\examples{
require(magrittr)
bsgenome <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38
gr <- GenomicRanges::GRanges(
          c(locus1 = 'chr11:5227002', locus2 = 'chr11:4699500'), 
            strand = c('-', '+'), 
            seqinfo = BSgenome::seqinfo(bsgenome))
gr \%>\%   up_flank(-22,  -1, plot = TRUE)
gr \%>\%   up_flank(-22,  -1, plot = TRUE, stranded = FALSE)
gr \%>\% down_flank( +1, +22, plot = TRUE)
gr \%>\% down_flank( +1, +22, plot = TRUE, stranded = FALSE)
gr \%>\%     extend(-10, +20, plot = TRUE)
gr \%>\%     extend(-10, +20, plot = TRUE, stranded = FALSE)
}
