% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_filter_doench_score.R
\name{add_contextseq}
\alias{add_contextseq}
\title{Add [-4, +3] contextseq}
\usage{
add_contextseq(cas9s, bsgenome, verbose = TRUE)
}
\arguments{
\item{cas9s}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{bsgenome}{\code{\link[BSgenome]{BSgenome-class}}}

\item{verbose}{logical(1)}
}
\value{
character vector
}
\description{
Add [-4, +3] contextseq
}
\examples{
require(magrittr)
bedfile <- system.file('extdata/SRF.bed', package = 'multicrispr')
txdb <- utils::getFromNamespace('TxDb.Mmusculus.UCSC.mm10.knownGene', 
                                'TxDb.Mmusculus.UCSC.mm10.knownGene')
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10                                
targets  <- bed_to_granges(bedfile, txdb)  \%>\% 
            double_flank() \%>\% 
            add_seq(bsgenome)
cas9s    <- find_cas9s(targets)
cas9s \%<>\% add_contextseq(bsgenome)
cas9s[1:3]$seq
cas9s[1:3]$contextseq
}
