% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_score_crispr_sites.R
\name{add_contextseq}
\alias{add_contextseq}
\title{Add [-4, +3] contextseq}
\usage{
add_contextseq(sites, bsgenome, verbose = TRUE)
}
\arguments{
\item{sites}{\code{\link[GenomicRanges]{GRanges-class}}: crispr sites}

\item{bsgenome}{\code{\link[BSgenome]{BSgenome-class}}}

\item{verbose}{logical(1)}
}
\value{
character vector
}
\description{
Add [-4, +3] contextseq
}
\examples{
require(magrittr)
bedfile <- system.file('extdata/SRF.bed', package = 'multicrispr')
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
targets  <- bed_to_granges(bedfile, 'mm10')  \%>\% 
            double_flank() \%>\% 
            add_seq(bsgenome)
sites    <- find_crispr_sites(targets)
sites \%<>\% add_contextseq(bsgenome)
sites[1:3]$seq
sites[1:3]$contextseq
}
