% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_filter_specific.R
\name{add_match_counts}
\alias{add_match_counts}
\alias{add_target_counts}
\alias{add_genome_counts}
\title{Add genome/targetset match counts}
\usage{
add_match_counts(
  spacers,
  indexdir,
  norc,
  mismatches = 2,
  outdir = OUTDIR,
  pam = "NGG",
  verbose = TRUE
)

add_target_counts(
  spacers,
  targets,
  bsgenome,
  mismatches = 2,
  pam = "NGG",
  outdir = OUTDIR,
  verbose = TRUE
)

add_genome_counts(
  spacers,
  bsgenome = getBSgenome(genome(spacers)[1]),
  mismatches = 2,
  pam = "NGG",
  outdir = OUTDIR,
  indexedgenomesdir = INDEXEDGENOMESDIR,
  verbose = TRUE
)
}
\arguments{
\item{spacers}{spacer \code{\link[GenomicRanges]{GRanges-class}}}

\item{mismatches}{number (default 2): max number of mismatches to consider}

\item{outdir}{dir where output is written to}

\item{pam}{string (default 'NGG') pam pattern to expand}

\item{verbose}{TRUE (default) or FALSE}

\item{targets}{target \code{\link[GenomicRanges]{GRanges-class}}}

\item{bsgenome}{\code{\link[BSgenome]{BSgenome-class}}}

\item{indexedgenomesdir}{string: dir with indexed genomes}
}
\value{
spacer GRanges with additional mcols
}
\description{
Count spacer matches to targetset/genome and add to GRanges
}
\details{
Expands iupac amgiguities in the pam sequence.
Matches all resulting sequences against (indexes) target and genome.
Adds match counts to GRanges object, and then returns it.
}
\examples{
# TFBS example
#-------------
 bs <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
 bedfile  <- system.file('extdata/SRF.bed', package = 'multicrispr')
 targets  <- extend(bed_to_granges(bedfile, genome = 'mm10'))
 spacers <- find_spacers(targets, bs)
 add_target_counts(spacers, targets, bs)
 add_match_counts( spacers, index_targets(targets, bs), norc=FALSE)
 # add_genome_counts(spacers, bs, indexedgenomesdir=index_genome(bs))

# PE example
#-----------
 require(magrittr)
 bs <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38  
 gr <- char_to_granges(c(PRNP = 'chr20:4699600:+',             # snp
                         HBB  = 'chr11:5227002:-',             # snp
                         HEXA = 'chr15:72346580-72346583:-',   # del
                         CFTR = 'chr7:117559593-117559595:+'), # ins
                       bs)
 spacers <- find_pe_spacers(gr, bs)
 # add_genome_counts(spacers, bs, indexedgenomesdir = index_genome(bs))
}
\seealso{
\code{\link{index_genome}}, \code{\link{index_targets}}
}
