% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_filter_specific.R
\name{add_genome_counts}
\alias{add_genome_counts}
\title{Add genome counts}
\usage{
add_genome_counts(
  spacers,
  bsgenome = getBSgenome(genome(spacers)[1]),
  mismatches = 2,
  pam = "NGG",
  outdir = OUTDIR,
  indexedgenomesdir = INDEXEDGENOMESDIR,
  plot = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{spacers}{spacer \code{\link[GenomicRanges]{GRanges-class}}}

\item{bsgenome}{\code{\link[BSgenome]{BSgenome-class}}}

\item{mismatches}{number (default 2): max number of mismatches to consider}

\item{pam}{string (default 'NGG') pam pattern to expand}

\item{outdir}{dir where output is written to}

\item{indexedgenomesdir}{string: dir with indexed genomes}

\item{plot}{FALSE (default) or TRUE}

\item{verbose}{TRUE (default) or FALSE}
}
\value{
spacer GRanges with additional mcols
}
\description{
Count spacer matches to genome and add to GRanges
}
\details{
Expands iupac amgiguities in the pam sequence.
Matches all resulting sequences against (indexes) target and genome.
Adds match counts to GRanges object, and then returns it.
}
\examples{
# PE example
#-----------
 require(magrittr)
 bsgenome <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38  
 gr <- char_to_granges(c(PRNP = 'chr20:4699600:+',             # snp
                         HBB  = 'chr11:5227002:-',             # snp
                         HEXA = 'chr15:72346580-72346583:-',   # del
                         CFTR = 'chr7:117559593-117559595:+'), # ins
                       bsgenome)
 spacers <- find_pe_spacers(gr, bsgenome)
 # index_genome(bsgenome)
 # add_genome_counts(spacers, bsgenome, mismatches=1, plot = TRUE)
    
# TFBS example
#-------------
 bsgenome <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
 bedfile  <- system.file('extdata/SRF.bed', package = 'multicrispr')
 targets <- extend(bed_to_granges(bedfile, genome = 'mm10'))
 spacers <- find_spacers(targets, bsgenome)
 # index_genome(bsgenome)
 # add_genome_counts(spacers, bsgenome)
 # add_genome_counts(spacers, bsgenome, mismatches=3)
}
\seealso{
\code{\link{index_genome}}, \code{\link{index_targets}}
}
