% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_find_cas9s.R
\name{add_inverse_strand}
\alias{add_inverse_strand}
\title{Add inverse strand for each range}
\usage{
add_inverse_strand(gr, plot = TRUE, verbose = TRUE)
}
\arguments{
\item{gr}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{plot}{logical(1)}

\item{verbose}{logical(1)}
}
\value{
\code{\link[GenomicRanges]{GRanges-class}}, twice as long as input
}
\description{
Add inverse strand for each range
}
\examples{
require(magrittr)
bedfile <- system.file('extdata/SRF.bed', package = 'multicrispr')
txdb <- utils::getFromNamespace(
             'TxDb.Mmusculus.UCSC.mm10.knownGene', 
             'TxDb.Mmusculus.UCSC.mm10.knownGene')
gr <- bed_to_granges(bedfile, txdb, plot = FALSE)
add_inverse_strand(gr)
}
