% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_extend_flank.R
\name{add_inverse_strand}
\alias{add_inverse_strand}
\title{Add inverse strand}
\usage{
add_inverse_strand(gr, plot = TRUE, verbose = TRUE)
}
\arguments{
\item{gr}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{plot}{TRUE (default) or FALSE}

\item{verbose}{TRUE (default) or FALSE}
}
\description{
Add inverse strand
}
\examples{
# Load
    require(magrittr)
    bsgenome <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38
    bsinfo <- BSgenome::seqinfo(bsgenome)
    
# PRNP snp: Kuru resistance variant (G -> T)
    gr <- GenomicRanges::GRanges(
           'chr20:4699500', strand = '+', seqinfo = bsinfo)
    gr \%<>\% add_seq(bsgenome))
    gr \%>\%  add_inverse_strand()
    
# HBB snp: sickle cell variant (T -> A)
    gr <- GenomicRanges::GRanges(
            'chr11:5227002-5227002', strand = '-', seqinfo = bsinfo)
    gr \%<>\% add_seq(bsgenome))
    gr \%>\%  add_inverse_strand()
    
# HEXA TATC duplication: Tay-Sachs variant
    gr <- GenomicRanges::GRanges(
            'chr15:72346580-72346583', strand = '-', seqinfo = bsinfo)
    gr \%<>\% add_seq(bsgenome))
    gr \%>\%  add_inverse_strand()
}
