% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_manipulate_ranges.R
\name{add_seq}
\alias{add_seq}
\title{Add sequence to GRanges}
\usage{
add_seq(gr, bsgenome, verbose = TRUE)
}
\arguments{
\item{gr}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{bsgenome}{\code{\link[BSgenome]{BSgenome-class}}}

\item{verbose}{logical(1)}
}
\value{
\code{\link[GenomicRanges]{GRanges-class}}
}
\description{
Add sequence to GRanges
}
\examples{
# SRF binding sites
    bedfile  <- system.file('extdata/SRF.bed', package = 'multicrispr')
    gr <- bed_to_granges(bedfile, 'mm10')
    bsgenome <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
    (gr <- add_seq(gr, bsgenome))
    
# PRNP snp: Kuru resistance variant (G -> T)
    require(magrittr)
    bsgenome <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38
    gr  <-  GRanges(
               'chr20:4699500', strand = '+', seqinfo = seqinfo(bsgenome))
    gr \%<>\% multicrispr::add_inverse_strand()
    gr \%<>\% multicrispr::extend(bsgenome = bsgenome)
   (gr \%<>\% multicrispr::add_seq(bsgenome))
}
