% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_filter_specific.R
\name{add_target_counts}
\alias{add_target_counts}
\title{Add target counts}
\usage{
add_target_counts(
  spacers,
  targets,
  bsgenome,
  outdir = default_outdir(),
  pam = "NGG",
  verbose = TRUE
)
}
\arguments{
\item{spacers}{spacer \code{\link[GenomicRanges]{GRanges-class}}}

\item{targets}{target \code{\link[GenomicRanges]{GRanges-class}}}

\item{bsgenome}{\code{\link[BSgenome]{BSgenome-class}}}

\item{outdir}{dir where output is written to}

\item{pam}{string (default 'NGG') pam pattern to expand}

\item{verbose}{TRUE (default) or FALSE}
}
\description{
Count spacer matches among targets
}
\details{
Expands iupac amgiguities in the pam sequence.
Matches all resulting sequences against (indexes) target and genome.
Adds match counts to GRanges object, and then returns it.
}
\examples{
# TFBS example
#-------------
 bsgenome <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
 genome <- '~/.multicrispr/bowtie/BSgenome.Mmusculus.UCSC.mm10'
 bedfile  <- system.file('extdata/SRF.bed', package = 'multicrispr')
 targets  <- extend(bed_to_granges(bedfile, genome = 'mm10'))
 spacers <- find_spacers(targets, bsgenome)
 spacers \%>\% add_target_counts(targets, bsgenome)
}
\seealso{
\code{\link{index_genome}}, \code{\link{index_targets}}
}
