% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_genes_to_granges.R
\name{annotate_granges}
\alias{annotate_granges}
\title{Annotate GRanges}
\usage{
annotate_granges(granges, db)
}
\arguments{
\item{granges}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{db}{\code{\link[GenomicFeatures]{TxDb-class}} or 
\code{\link[ensembldb]{EnsDb-class}}}
}
\value{
\code{\link[GenomicRanges]{GRanges-class}}
}
\description{
Annotate GRanges
}
\examples{
# Read 
    bedfile <- system.file('extdata/SRF.bed', package = 'multicrispr')
    granges <- bed_to_granges(bedfile, 'mm10', plot = FALSE)
    
# Annotate Entrez 
    dbname <- 'TxDb.Mmusculus.UCSC.mm10.knownGene'
    db <- utils::getFromNamespace(dbname, dbname)
    annotate_granges(granges, db)

# Annotate Ensembl
    db <- EnsDb.Mmusculus.v98()
    annotate_granges(granges, db)
}
\seealso{
\code{\link{genes_to_granges}} (inverse operation)
}
