% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_read_assert.R
\name{as.granges}
\alias{as.granges}
\title{Convert data.table into GenomicRanges::GRanges}
\usage{
as.granges(x, bsgenome)
}
\arguments{
\item{x}{data.table}

\item{bsgenome}{BSgenome, e.g. BSgenome.Mmusculus.UCSC.mm10::Mmusculus}
}
\value{
GenomicRanges::GRanges
}
\description{
Convert data.table into GenomicRanges::GRanges
}
\examples{
require(magrittr)
bedfile <- system.file('extdata/SRF_sites.bed', package = 'crisprapex')
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
granges <- read_bed(bedfile, bsgenome)
granges$seqs <- seqs(granges)
x <- granges \%>\% data.table::as.data.table() 
x \%>\% as.granges(bsgenome = bsgenome)
}
