% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_bed_to_granges.R
\name{bed_to_granges}
\alias{bed_to_granges}
\title{Read bedfile into GRanges}
\usage{
bed_to_granges(bedfile, genome = NULL, do_order = TRUE, plot = TRUE,
  verbose = TRUE)
}
\arguments{
\item{bedfile}{file path}

\item{genome}{genome identifier ('mm10')}

\item{do_order}{logical(1)}

\item{plot}{logical(1)}

\item{verbose}{logical(1)}
}
\value{
\code{\link[GenomicRanges]{GRanges-class}}
}
\description{
Read bedfile into granges and output graphical and textual summaries
}
\examples{
bedfile  <- system.file('extdata/SRF.bed', package = 'multicrispr')
bed_to_granges(bedfile, 'mm10')
}
\seealso{
\code{rtracklayer::import.bed} (documented in 
\code{\link[rtracklayer]{BEDFile-class}}), around which this function wraps.
}
