% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utilities.R
\name{canonicalseqlevels}
\alias{canonicalseqlevels}
\title{Get canonical seqlevels}
\usage{
canonicalseqlevels(gr)
}
\arguments{
\item{gr}{\code{\link[GenomicRanges]{GRanges-class}}}
}
\value{
character vector
}
\description{
Get canonical seqlevels
}
\examples{

# Mouse entrez
    dbname <- 'TxDb.Mmusculus.UCSC.mm10.knownGene'
    db <- utils::getFromNamespace(dbname, dbname)
    gr <- GenomicFeatures::genes(db)
    GenomeInfoDb::seqlevels(gr)
    canonicalseqlevels(gr)

# Human entrez
    dbname <- 'TxDb.Hsapiens.UCSC.hg38.knownGene'
    db <- utils::getFromNamespace(dbname, dbname)
    gr <- GenomicFeatures::genes(db)
    GenomeInfoDb::seqlevels(gr)
    canonicalseqlevels(gr)
    
# Mouse ensembl
    db <- EnsDb.Mmusculus.v98()
    gr <- ensembldb::genes(db)
    GenomeInfoDb::seqlevels(gr)
    canonicalseqlevels(gr)
    
# Human ensembl
    db <- EnsDb.Hsapiens.v98()
    gr <- ensembldb::genes(db)
    GenomeInfoDb::seqlevels(gr)
    canonicalseqlevels(gr)
    
# Mouse Bedfile
    bedfile <- system.file('extdata/SRF.bed', package = 'multicrispr')
    gr <- read_bed(bedfile, 'mm10', plot = FALSE)
    GenomeInfoDb::seqlevels(gr)
    canonicalseqlevels(gr)
    
}
\seealso{
\code{\link[GenomeInfoDb]{seqinfo}}
}
