% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_read_flank_slop.R
\name{complement}
\alias{complement}
\title{Complement ranges}
\usage{
complement(ranges, verbose = TRUE)
}
\arguments{
\item{ranges}{data.table(chr, start, end, strand)}

\item{verbose}{logical(1)}
}
\value{
data.table(chr, start, end, strand)
}
\description{
Complement ranges
}
\examples{
bedfile <- system.file('extdata/SRF_sites.bed', package = 'crisprapex') 
ranges <- read_bed(bedfile)
complement(ranges) 
}
