% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_flank_slop.R
\name{complement}
\alias{complement}
\title{Complement granges}
\usage{
complement(granges, verbose = TRUE)
}
\arguments{
\item{granges}{GenomicRanges::GRanges}

\item{verbose}{logical(1)}
}
\value{
GenomicRanges::GRanges
}
\description{
Complement granges
}
\examples{
bedfile <- system.file('extdata/SRF_sites.bed', package = 'crisprapex') 
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
granges <- read_bed(bedfile, bsgenome)
granges
complement(granges) 
}
