% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_score_cas9s.R
\name{contextify_start}
\alias{contextify_start}
\alias{contextify_end}
\title{Contextify start/stop}
\usage{
contextify_start(chr, start, strand)

contextify_end(chr, end, strand)
}
\arguments{
\item{chr}{character vector: chromosome values}

\item{start}{numeric vector: start positions}

\item{strand}{character vector: '+' or '-' values}

\item{end}{numeric vector: end positions}
}
\value{
numeric vector
}
\description{
Contextify start/stop
}
\examples{
ranges <- data.table::data.table(
            chr    = c('chr1', 'chr1'), 
            start  = c(200, 300), 
            end    = c(220, 330), 
            strand = c('-', '+'))
ranges [ , contextstart := contextify_start(chr, start, end, strand) ][]
ranges [ , contextend   := contextify_end(  chr, start, end, strand) ][]
}
