% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_score_cas9s.R
\name{contextseqs}
\alias{contextseqs}
\title{Get contextseqs}
\usage{
contextseqs(cas9ranges)
}
\arguments{
\item{cas9ranges}{\code{\link[GenomicRanges]{GRanges-class}}}
}
\value{
character vector
}
\description{
Get [-3, +3] contextseqs for given cas9ranges
}
\examples{
require(magrittr)
bedfile <- system.file('extdata/SRF.bed', package = 'multicrispr')
cas9ranges  <-  read_bed(bedfile, 'mm10') \%>\% 
                flank_fourways() \%>\% 
                find_cas9ranges()
cas9ranges[1:3] \%>\% seqs()
cas9ranges[1:3] \%>\% contextseqs()
}
