% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_score_cas9s.R
\name{contextseqs}
\alias{contextseqs}
\title{Get contextseqs}
\usage{
contextseqs(cas9ranges)
}
\arguments{
\item{cas9ranges}{GenomicRanges::GRanges}
}
\description{
Get [-3, +3] contextseqs for given cas9ranges
}
\examples{
require(magrittr)
bedfile <- system.file('extdata/SRF_sites.bed', package = 'crisprapex')
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
cas9ranges  <-  read_bed(bedfile, bsgenome) \%>\% 
                flank_fourways() \%>\% 
                find_cas9s()
cas9ranges[1:10] \%>\% contextseqs()
}
