% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_filter_offtargetfree_sites.R
\name{count_target_matches}
\alias{count_target_matches}
\alias{count_genome_matches}
\title{Count target/genome matches}
\usage{
count_target_matches(crisprseqs, targetseqs, mismatch, verbose = TRUE)

count_genome_matches(
  crisprseqs,
  bsgenome,
  mismatch,
  chromosomes = BSgenome::seqnames(bsgenome),
  verbose = TRUE
)
}
\arguments{
\item{crisprseqs}{character() or \code{\link[Biostrings]{XStringSet-class}}}

\item{targetseqs}{character() or \code{\link[Biostrings]{XStringSet-class}}}

\item{mismatch}{number: number of allowed mismatches}

\item{verbose}{logical(1)}

\item{bsgenome}{\code{\link[BSgenome]{BSgenome-class}}}

\item{chromosomes}{character vector}
}
\value{
numeric(length(crisprseqs))
}
\description{
Count target/genome matches
}
\examples{
# Read ranges and extend
    require(magrittr)
    bedfile <- system.file('extdata/SRF.bed', package = 'multicrispr')
    targets <- bed_to_granges(bedfile, 'mm10')
    targets \%<>\% extend()

# Add seqs and find crispr sites
    bsgenome <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
    targets \%<>\% add_seq(bsgenome)
    sites <- find_crispr_sites(targets)

# Count matches
    crisprseqs <- sites$seq[1:10]
    count_target_matches(crisprseqs, targets$seq, 0)
    count_genome_matches(crisprseqs, bsgenome,    0, chromosomes = 'chrY')
}
