% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_find_cas9s.R
\name{count_target_matches}
\alias{count_target_matches}
\title{Count target matches}
\usage{
count_target_matches(cas9seqs, targetseqs, mismatch, verbose = TRUE)
}
\arguments{
\item{cas9seqs}{character() or \code{\link[Biostrings]{XStringSet-class}}}

\item{targetseqs}{character() or \code{\link[Biostrings]{XStringSet-class}}}

\item{mismatch}{number: number of allowed mismatches}

\item{verbose}{logical(1)}
}
\value{
numeric(length(cas9seqs))
}
\description{
Count target matches
}
\examples{

# Read target ranges
require(magrittr)
bedfile <- system.file('extdata/SRF.bed', package = 'multicrispr')
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
targetranges <- read_bed(bedfile, bsgenome)
targetranges \%<>\% slop_fourways \%>\% extract(1:10)

# Find cas9ranges
cas9ranges <- find_cas9ranges(targetranges)

# Count target matches
count_target_matches(
    cas9seqs   = getSeq(bsgenome, cas9ranges), 
    targetseqs = getSeq(bsgenome, targetranges), 
    mismatch   = 0, 
    verbose    = TRUE)
}
\seealso{
\code{\link{count_target_matches}}, \code{\link{vcountPDict}}
}
