% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_filter_no_offtargets.R
\name{count_target_matches}
\alias{count_target_matches}
\alias{count_genome_matches}
\title{Count target/genome matches}
\usage{
count_target_matches(cas9seqs, targetseqs, mismatch, verbose = TRUE)

count_genome_matches(cas9seqs, bsgenome, mismatch,
  chromosomes = seqnames(bsgenome), verbose = TRUE)
}
\arguments{
\item{cas9seqs}{character() or \code{\link[Biostrings]{XStringSet-class}}}

\item{targetseqs}{character() or \code{\link[Biostrings]{XStringSet-class}}}

\item{mismatch}{number: number of allowed mismatches}

\item{verbose}{logical(1)}

\item{bsgenome}{\code{\link[BSgenome]{BSgenome-class}}}

\item{chromosomes}{character vector}
}
\value{
numeric(length(cas9seqs))
}
\description{
Count target/genome matches
}
\examples{
# Read ranges and extend
    require(magrittr)
    bedfile <- system.file('extdata/SRF.bed', package = 'multicrispr')
    txdb <-  utils::getFromNamespace( 'TxDb.Mmusculus.UCSC.mm10.knownGene', 
                                      'TxDb.Mmusculus.UCSC.mm10.knownGene')
    targets <- bed_to_granges(bedfile, txdb)
    targets \%<>\% extend()

# Add seqs and find cas9s
    bsgenome <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
    targets \%<>\% add_seq(bsgenome)
    cas9s <- find_cas9s(targets)

# Count matches
    cas9sseqs <- cas9s$seq[1:10]
    count_target_matches(cas9seqs, targets$seq, 0)
    count_genome_matches(cas9seqs, bsgenome,    0, chromosomes = 'chrY')
}
