% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_manipulate_ranges.R
\name{double_flank}
\alias{double_flank}
\title{Double flank}
\usage{
double_flank(
  gr,
  upstart = -200,
  upend = -1,
  downstart = 1,
  downend = 200,
  do_reduce = TRUE,
  add_inverse = TRUE
)
}
\arguments{
\item{gr}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{upstart}{upstream flank start in relation to start(gr)}

\item{upend}{upstream flank end   in relation to start(gr)}

\item{downstart}{downstream flank start in relation to end(gr)}

\item{downend}{downstream flank end   in relation to end(gr)}

\item{do_reduce}{TRUE(default) or FALSE: merge overlapping ranges?}

\item{add_inverse}{TRUE (default) or FALSE: add inverse strand?}
}
\value{
\code{\link[GenomicRanges]{GRanges-class}}
}
\description{
Double flanks, reduce overlaps, add reverse complement
}
\details{
This function is created for crispr in the context of crisprapex.
It converts a set of crisprapex targets into a set of ranges in which to 
search for crispr spacers. This involves (1) agnostifying the strand, 
(2) calcuting up and down flanks, (3) merging overlapping ranges, and 
(4) extending each range to both strands
}
\examples{
gr  <-  bed_to_granges(bedfile, genome = 'mm10')
double_flank(gr)
}
