% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_extend_flank.R
\name{double_flank}
\alias{double_flank}
\title{Double flank}
\usage{
double_flank(gr, leftstart = -200, leftend = -1, rightstart = 1,
  rightend = 200, complement = TRUE, plot = TRUE, verbose = TRUE)
}
\arguments{
\item{gr}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{leftstart}{number: left flank start  (relative to range start)}

\item{leftend}{number: left flank  end   (relative to range start)}

\item{rightstart}{number: right flank start (relative to range end)}

\item{rightend}{number: right flank end   (relative to range end)}

\item{complement}{logical(1): whether to add complementary strand}

\item{plot}{logical(1)}

\item{verbose}{logical(1)}
}
\value{
\code{\link[GenomicRanges]{GRanges-class}}
}
\description{
Flank left and right and merge overlaps
}
\examples{
bedfile <- system.file('extdata/SRF.bed', package = 'multicrispr')
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
gr <- bed_to_granges(bedfile, bsgenome, plot = FALSE)
double_flank(gr)
}
