% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utilities.R
\name{dt2gr}
\alias{dt2gr}
\alias{gr2dt}
\title{Switch between data.table and GenomicRanges::GRanges representation}
\usage{
dt2gr(dt)

gr2dt(gr)
}
\arguments{
\item{dt}{data.table}

\item{gr}{GenomicRanges::GRanges}
}
\value{
IRanges
}
\description{
Switch between data.table and GenomicRanges::GRanges representation
}
\examples{
require(magrittr) 
dt <- data.table::data.table(
    chr    = 'chr1',
    start  = c( 1,  5, 5), 
    end    = c(10, 15, 15), 
    strand = c('+', '+', '-')
)
dt \%>\% dt2gr()
dt \%>\% dt2gr() \%>\% gr2dt()
}
