% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_data.R
\name{EnsDb.Hsapiens.v98}
\alias{EnsDb.Hsapiens.v98}
\title{Get EnsDb.Hsapiens.v98 from AnnotationHub}
\usage{
EnsDb.Hsapiens.v98()
}
\value{
\code{\link[ensembldb]{EnsDb-class}}
}
\description{
Get EnsDb.Hsapiens.v98 from AnnotationHub
}
\examples{
EnsDb.Hsapiens.v98()
}
