% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_filter_specific_cas9s.R
\name{filter_no_offtargets}
\alias{filter_no_offtargets}
\title{Filter for specific (i.e. offtarget-free) cas9 sites}
\usage{
filter_no_offtargets(cas9ranges, targetranges, mismatch = 2,
  offtargetchr = canonicalseqlevels(targetranges), verbose = TRUE)
}
\arguments{
\item{cas9ranges}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{targetranges}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{mismatch}{number: max number of mismatches to consider}

\item{offtargetchr}{character vector: chromosomes for offtarget analysis, 
probably generated with genomeInfoDb::seqlevels(targetranges) or 
canonicalseqlevels(targetranges)}

\item{verbose}{logical(1)}
}
\value{
\code{\link[GenomicRanges]{GRanges-class}}
        mcols(GRanges) contains sequences and match counts:
            matches0 = perfect match counts
            matches1 = single mismatch counts
            matches2 = double mismatch counts
}
\description{
Filter for specific (i.e. offtarget-free) cas9 sites
}
\examples{
# Note: restricting example to 'chrY' only to keep it fast
require(magrittr)
bedfile  <- system.file('extdata/SRF.bed', package = 'multicrispr')
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
targetranges <- bed_to_granges(bedfile, bsgenome, plot = FALSE) \%>\% 
                extract(GenomeInfoDb::seqnames(.)=='chrY') \%>\% 
                extend()
find_offtargetfree_cas9s(targetranges, 0, offtargetchr = 'chrY')
}
