% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_filter_specific.R
\name{filter_prime_specific}
\alias{filter_prime_specific}
\title{Filter for specific prime editing spacers}
\usage{
filter_prime_specific(
  spacers,
  indexedgenome = default_indexedgenome(spacers),
  mismatches = 2,
  outdir = default_outdir(),
  pam = "NGG",
  verbose = TRUE
)
}
\arguments{
\item{spacers}{spacer \code{\link[GenomicRanges]{GRanges-class}}}

\item{indexedgenome}{bowtie indexed genome dir}

\item{outdir}{directory where output is written to}

\item{pam}{string (default 'NGG'): pam sequence}

\item{verbose}{TRUE (default) or FALSE}
}
\description{
Filters spacers which are specific for prime editing site
}
\examples{
# PE example
#-----------
 require(magrittr)
 bsgenome <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38  
 gr <- char_to_granges(c(PRNP = 'chr20:4699600:+',             # snp
                         HBB  = 'chr11:5227002:-',             # snp
                         HEXA = 'chr15:72346580-72346583:-',   # del
                         CFTR = 'chr7:117559593-117559595:+'), # ins
                       bsgenome)
 spacers <- find_pe_spacers(gr, bsgenome)
 indexedgenome <- '~/.multicrispr/bowtie/BSgenome.Hsapiens.UCSC.hg38'
 # index_genome(bsgenome, indexedgenome) # one time effort - takes few h
 # filter_prime_specific(spacers, indexedgenome)
}
