% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_add_genome_matches.R
\name{filter_target_specific}
\alias{filter_target_specific}
\title{Filter for target-specific spacers}
\usage{
filter_target_specific(
  spacers,
  targets,
  bsgenome,
  indexedgenome,
  outdir = default_outdir(),
  pam = "NGG",
  verbose = TRUE
)
}
\arguments{
\item{spacers}{spacer \code{\link[GenomicRanges]{GRanges-class}}}

\item{targets}{target \code{\link[GenomicRanges]{GRanges-class}}}

\item{indexedgenome}{bowtie indexed genome dir}

\item{outdir}{directory where output is written to}

\item{pam}{string (default 'NGG'): pam sequence}

\item{verbose}{TRUE (default) or FALSE}
}
\description{
Filter for target-specific spacers
}
\examples{
# TFBS example
#-------------
 bsgenome <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
 bedfile  <- system.file('extdata/SRF.bed', package = 'multicrispr')
 targets <- extend(bed_to_granges(bedfile, genome = 'mm10'))
 spacers <- find_spacers(targets, bsgenome)
 indexedgenome <- '~/.multicrispr/bowtie/BSgenome.Mmusculus.UCSC.mm10'
 # index_genome(indexedgenome)
 # filter_target_specific(spacers, targets, bsgenome, indexedgenome)
}
