% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_find_cas9s.R
\name{find_cas9ranges}
\alias{find_cas9ranges}
\title{Find cas9 ranges in targetranges}
\usage{
find_cas9ranges(targetranges, verbose = TRUE)
}
\arguments{
\item{targetranges}{GenomicRanges::GRanges}

\item{verbose}{logical(1)}

\item{specific}{logical(1)}
}
\value{
GenomicRanges::GRanges
}
\description{
Find cas9 ranges in targetranges
}
\examples{
require(magrittr)
bedfile  <- system.file('extdata/SRF_sites.bed', package='crisprapex')
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
targetranges <- read_bed(bedfile, bsgenome) \%>\% flank_fourways()
find_cas9ranges(targetranges)
}
