% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_find_cas9s.R
\name{find_cas9s}
\alias{find_cas9s}
\title{Find cas9 sites in targetranges}
\usage{
find_cas9s(gr, inclcompl = TRUE, verbose = TRUE)
}
\arguments{
\item{gr}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{inclcompl}{logical(1): include complementary strands in search?}

\item{verbose}{logical(1): report verbosely?}
}
\value{
\code{\link[GenomicRanges]{GRanges-class}}
}
\description{
Find cas9 sites in targetranges
}
\examples{
bedfile  <- system.file('extdata/SRF.bed', package='multicrispr')
gr <- extend(read_bed(bedfile, 'mm10'), plot = FALSE)
find_cas9s(gr)
}
