% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_find_cas9s.R
\name{find_cas9s}
\alias{find_cas9s}
\title{Find cas9 sites in targetranges}
\usage{
find_cas9s(targets, plot = TRUE, verbose = TRUE)
}
\arguments{
\item{targets}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{plot}{TRUE (default) or FALSE}

\item{verbose}{TRUE (default) or FALSE}
}
\value{
\code{\link[GenomicRanges]{GRanges-class}}
}
\description{
Find cas9 sites in targetranges
}
\examples{
# Read bed into granges, extend, add seqs
    require(magrittr)
    bsgenome <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
    bedfile  <- system.file('extdata/SRF.bed', package='multicrispr')
    targets <- bed_to_granges(bedfile, 'mm10')
    targets \%<>\% extend(plot = FALSE)
    targets \%<>\% add_seq(bsgenome)
    
# Find cas9 seqs
    find_cas9s(targets)
}
