% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_find_cas9s.R
\name{find_cas9s}
\alias{find_cas9s}
\title{Find cas9 sites in targetranges}
\usage{
find_cas9s(gr, bsgenome, inclcompl = TRUE, verbose = TRUE)
}
\arguments{
\item{gr}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{bsgenome}{\code{\link[BSgenome]{BSgenome-class}}}

\item{inclcompl}{logical(1): include complementary strands in search?}

\item{verbose}{logical(1): report verbosely?}
}
\value{
\code{\link[GenomicRanges]{GRanges-class}}
}
\description{
Find cas9 sites in targetranges
}
\examples{
bedfile  <- system.file('extdata/SRF.bed', package='multicrispr')
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
gr <- extend(bed_to_granges(bedfile, bsgenome), plot = FALSE)
find_cas9s(gr)
}
