% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_find_cas9s.R
\name{find_cas9s}
\alias{find_cas9s}
\title{Find cas9 sites in given ranges}
\usage{
find_cas9s(ranges, bsgenome, verbose = TRUE)
}
\arguments{
\item{ranges}{data.table(chr, start, end, strand)}

\item{bsgenome}{BSgenome object (eg BSgenome.Mmusculus.UCSC.mm10::Mmusculus)}

\item{verbose}{logical(1)}
}
\value{
data.table(chr, start, end, strand, cas9seq, cas9start, cas9end). \cr
        One row per cas9 site
}
\description{
Find cas9 sites in given ranges
}
\examples{
require(magrittr)
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
bedfile  <- system.file('extdata/SRF_sites.bed', package='crisprapex')
read_bed(bedfile) \%>\% slop_fourways(bsgenome) \%>\% find_cas9s(bsgenome)
}
