% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_find_cas9s.R
\name{find_cas9s}
\alias{find_cas9s}
\title{Find cas9 sites in targetranges}
\usage{
find_cas9s(gr, inclcompl = TRUE, plot = TRUE, verbose = TRUE)
}
\arguments{
\item{gr}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{inclcompl}{logical(1): include complementary strands in search?}

\item{plot}{logical(1)}

\item{verbose}{logical(1)}
}
\value{
\code{\link[GenomicRanges]{GRanges-class}}
}
\description{
Find cas9 sites in targetranges
}
\examples{
# Read bed into granges, extend, add seqs
    bedfile  <- system.file('extdata/SRF.bed', package='multicrispr')
    txdb <- utils::getFromNamespace(
             'TxDb.Mmusculus.UCSC.mm10.knownGene', 
             'TxDb.Mmusculus.UCSC.mm10.knownGene')
    gr <- extend(bed_to_granges(bedfile, txdb), plot = FALSE)
    gr <- add_seq(gr, bsgenome)
    
# Find cas9 seqs
    find_cas9s(gr)
}
