% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_find_crispr_sites.R
\name{find_crispr_sites}
\alias{find_crispr_sites}
\title{Find crispr sites in targetranges}
\usage{
find_crispr_sites(
  gr,
  bsgenome,
  spacer = strrep("N", 20),
  pam = "NGG",
  plot = TRUE
)
}
\arguments{
\item{gr}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{bsgenome}{\code{\link[BSgenome]{BSgenome-class}}}

\item{spacer}{string: spacer pattern in extended IUPAC alphabet}

\item{pam}{string: pam pattern in extended IUPAC alphabet}

\item{plot}{TRUE (default) or FALSE}
}
\value{
\code{\link[GenomicRanges]{GRanges-class}}
}
\description{
Find crispr sites in targetranges
}
\examples{
# Small GRanges
    require(magrittr)
    bsgenome <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38  
    gr <- GenomicRanges::GRanges(
              seqnames = c(PRNP = 'chr20:4699600',             # snp
                           HBB  = 'chr11:5227002',             # snp
                           HEXA = 'chr15:72346580-72346583',   # del
                           CFTR = 'chr7:117559593-117559595'), # ins
              strand   = c(PRNP = '+', HBB = '-', HEXA = '-', CFTR = '+'), 
              seqinfo  = BSgenome::seqinfo(bsgenome))
    gr \%<>\% extend_for_pe()
    find_crispr_sites(gr, bsgenome)
    
# Large GRanges
    require(magrittr)
    bsgenome <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
    bedfile  <- system.file('extdata/SRF.bed', package='multicrispr')
    gr <- bed_to_granges(bedfile, 'mm10')
    gr \%<>\% extend(plot = FALSE)
    find_crispr_sites(gr, bsgenome)
}
\seealso{
\code{\link{find_prime_sites}} to find prime editing sites
}
