% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_find_crispr_sites.R
\name{find_crispr_sites}
\alias{find_crispr_sites}
\title{Find crispr sites in targetranges}
\usage{
find_crispr_sites(targets, plot = TRUE, verbose = TRUE)
}
\arguments{
\item{targets}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{plot}{TRUE (default) or FALSE}

\item{verbose}{TRUE (default) or FALSE}
}
\value{
\code{\link[GenomicRanges]{GRanges-class}}
}
\description{
Find crispr sites in targetranges
}
\examples{
# Read bed into granges, extend, add seqs
    require(magrittr)
    bsgenome <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
    bedfile  <- system.file('extdata/SRF.bed', package='multicrispr')
    targets <- bed_to_granges(bedfile, 'mm10')
    targets \%<>\% extend(plot = FALSE)
    targets \%<>\% add_seq(bsgenome)
    
# Find crispr sites
    find_crispr_sites(targets)
}
