% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_find_pe_spacers.R
\name{find_pe_spacers}
\alias{find_pe_spacers}
\title{Find prime editing spacers}
\usage{
find_pe_spacers(
  gr,
  bsgenome,
  fixes = get_plus_seq(bsgenome, gr),
  nprimer = 13,
  nrt = 16,
  plot = TRUE
)
}
\arguments{
\item{gr}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{bsgenome}{\code{\link[BSgenome]{BSgenome-class}}}

\item{fixes}{character vector: '+' strand fix seqs}

\item{nprimer}{n primer nucleotides (default 13, max 17)}

\item{nrt}{n rev transcr nucleotides (default 16, recomm. 10-16)}

\item{plot}{TRUE (default) or FALSE}
}
\value{
\code{\link[GenomicRanges]{GRanges-class}} with PE spacer ranges
Each prime editing range is defined in terms of its N20NGG spacer.
Additionally, three sequence mcols are returned:
  * spacer: N20 spacers
  * pam:    NGG PAMs
  * extension: 3' extension of gRNA (RTtemplate + primerbindingsite)
}
\description{
Find prime editing spacers around target ranges
}
\details{
Below the architecture of a prime editing site.
Fixes can be performed anywhere in the revtranscript area.

        spacer        pam
  --------------------===
          primer     revtranscript
      -------------================
  1..............17....GG..........
  .....................CC..........
      ----------extension----------
}
\examples{
# PE example
#-----------
    require(magrittr)
    bsgenome <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38  
    gr <- GenomicRanges::GRanges(
              seqnames = c(PRNP = 'chr20:4699600',             # snp
                           HBB  = 'chr11:5227002',             # snp
                           HEXA = 'chr15:72346580-72346583',   # del
                           CFTR = 'chr7:117559593-117559595'), # ins
              strand   = c(PRNP = '+', HBB = '-', HEXA = '-', CFTR = '+'), 
              seqinfo  = BSgenome::seqinfo(bsgenome))
    find_pe_spacers(gr, bsgenome)
    find_spacers(extend_for_pe(gr), bsgenome)
}
\seealso{
\code{\link{find_spacers}} to find standard crispr sites
}
