% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_find_pe_spacers.R
\name{find_pe_spacers}
\alias{find_pe_spacers}
\title{Find prime editing spacers}
\usage{
find_pe_spacers(
  gr,
  bsgenome,
  edits = get_plus_seq(bsgenome, gr),
  nprimer = 13,
  nrt = 16,
  ontargets = "Doench2014",
  plot = TRUE,
  outdir = OUTDIR,
  indexedgenomesdir = INDEXEDGENOMESDIR,
  ...
)
}
\arguments{
\item{gr}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{bsgenome}{\code{\link[BSgenome]{BSgenome-class}}}

\item{edits}{character vector: desired edits on '+' strand.
If named, names should be identical to those of \code{gr}}

\item{nprimer}{n primer nucleotides (default 13, max 17)}

\item{nrt}{n rev transcr nucleotides (default 16, recomm. 10-16)}

\item{ontargets}{'Doench2014' or 'Doench2016': on-target scoring method}

\item{plot}{TRUE (default) or FALSE}

\item{outdir}{string: passed to filter_offtargets}

\item{indexedgenomesdir}{string: passed to filter_offtargets}

\item{...}{passed to plot_intervals}
}
\value{
\code{\link[GenomicRanges]{GRanges-class}} with PE spacer ranges
Each prime editing range is defined in terms of its N20NGG spacer.
Additionally, three sequence mcols are returned:
  * crisprspacer: N20 spacers
  * crisprpam:    NGG PAMs
  * crisprprimer: primer (on PAM strand)
  * crisprtranscript: reverse transcript (on PAM strand)
  * crisprextension:  3' extension of gRNA 
                (reverse transcription template + primer binding site)
}
\description{
Find prime editing spacers around target ranges
}
\details{
Below the architecture of a prime editing site.
Edits can be performed anywhere in the revtranscript area.

        spacer        pam
  --------------------===
          primer     revtranscript
      -------------================
  1..............17....GG..........
  .....................CC..........
      ----------extension----------
}
\examples{
# Find PE spacers for 4 clinically relevant loci (Anzalone et al, 2019)
    bsgenome <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38  
    gr <- char_to_granges(c(
        PRNP = 'chr20:4699600:+',             # snp: prion disease
        HBB  = 'chr11:5227002:-',             # snp: sickle cell anemia
        HEXA = 'chr15:72346580-72346583:-',   # del: tay sachs disease
        CFTR = 'chr7:117559593-117559595:+'), # ins: cystic fibrosis
        bsgenome)
    spacers <- find_pe_spacers(gr, bsgenome)
    spacers <- find_spacers(extend_for_pe(gr), bsgenome, complement = FALSE)
    
# Edit PRNP locus for resistance against prion disease (Anzalone et al, 2019)
    bsgenome <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38  
    gr <- char_to_granges('chr20:4699600:+', bsgenome)
    find_pe_spacers(gr, bsgenome)
    find_pe_spacers(gr, bsgenome, edits = 'T')
}
\seealso{
\code{\link{find_spacers}} to find standard crispr sites
}
