% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_find_cas9s.R
\name{find_specific_cas9ranges}
\alias{find_specific_cas9ranges}
\title{Find target cas9 ranges with no offtargets}
\usage{
find_specific_cas9ranges(targetranges, mismatch = 2, verbose = TRUE)
}
\arguments{
\item{targetranges}{GenomicRanges::GRanges}

\item{mismatch}{max number of mismatches to consider}

\item{verbose}{logical(1)}
}
\value{
GenomicRanges::GRanges
        mcols(GRanges) contains sequences and match counts:
            matches0 = perfect match counts
            matches1 = single mismatch counts
            matches2 = double mismatch counts
}
\description{
Find target cas9 ranges with no offtargets
}
\examples{
\\donrun{
   require(magrittr)
   bedfile <- system.file('extdata/SRF_sites.bed', package = 'multicrispr')
   bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
   targetranges <- read_bed(bedfile, bsgenome)[1:10] \%>\% flank_fourways()
   targetranges \%>\% find_specific_cas9ranges(mismatch=0)
}
}
