% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_find_cas9s.R
\name{find_specific_cas9s}
\alias{find_specific_cas9s}
\title{Find target-specific cas9seqs}
\usage{
find_specific_cas9s(targetranges, verbose = TRUE)
}
\arguments{
\item{targetranges}{GenomicRanges::GRanges}

\item{verbose}{logical(1)}
}
\value{
data.table(cas9seq, 
                   matches0, 
                   matches1, 
                   matches2, 
                   seqnames, 
                   start, 
                   end, 
                   width, 
                   strand)
}
\description{
Find target-specific cas9seqs
}
\examples{
\\donrun{
   require(magrittr)
   bedfile <- system.file('extdata/SRF_sites.bed', package = 'crisprapex')
   bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
   targetranges <- read_bed(bedfile, bsgenome)[1:10] \%>\% flank_fourways()
   targetranges \%>\% find_specific_cas9s()
}
}
