% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_find_cas9sites.R
\name{findcas9s}
\alias{findcas9s}
\title{Find cas9 sites in given ranges}
\usage{
findcas9s(ranges, bsgenome, verbose = TRUE)
}
\arguments{
\item{ranges}{data.table(chr, start, end, strand)}

\item{bsgenome}{BSgenome object (eg BSgenome.Mmusculus.UCSC.mm10::Mmusculus)}
}
\value{
data.table(chr, start, end, strand, cas9seq, cas9start, cas9end). \cr
        One row per cas9 site
}
\description{
Find cas9 sites in given ranges
}
\examples{
require(magrittr)
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
bedfile  <- system.file('extdata/SRF_sites.bed', package='cas9tbflanks')
read_bed(bedfile) \%>\% slop_fourways() \%>\% findcas9s(bsgenome)
}
