% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_flank_slop.R
\name{flank_fourways}
\alias{flank_fourways}
\title{Flank fourways}
\usage{
flank_fourways(gr, leftstart = -200, leftend = -1, rightstart = 1,
  rightend = 200, plot = TRUE, verbose = TRUE)
}
\arguments{
\item{gr}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{leftstart}{number: left flank start  (relative to range start)}

\item{leftend}{number: left flank  end   (relative to range start)}

\item{rightstart}{number: right flank start (relative to range end)}

\item{rightend}{number: right flank end   (relative to range end)}

\item{plot}{logical(1)}

\item{verbose}{logical(1)}
}
\value{
\code{\link[GenomicRanges]{GRanges-class}}
}
\description{
Flank left and right, for both strands, and merge overlaps
}
\examples{
bedfile <- system.file('extdata/SRF.bed', package = 'multicrispr')
gr <- read_bed(bedfile, 'mm10', plot = FALSE)
flank_fourways(gr)
}
