% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_read_flank_slop.R
\name{flank_fourways}
\alias{flank_fourways}
\title{Flank left/right ranges for both strands, merging overlaps}
\usage{
flank_fourways(ranges, leftstartoffset = -200, leftendoffset = -1,
  rightstartoffset = 1, rightendoffset = 200, verbose = TRUE)
}
\arguments{
\item{ranges}{data.table(chr, start, end, strand)}

\item{leftstartoffset}{numeric(1)}

\item{leftendoffset}{numeric(1)}

\item{rightstartoffset}{numeric(1)}

\item{rightendoffset}{numeric(1)}

\item{verbose}{logical(1): report?}
}
\value{
data.table(chr, start, end, strand)
}
\description{
Flank left/right ranges for both strands, merging overlaps
}
\examples{
bedfile <- system.file('extdata/SRF_sites.bed', package = 'cas9tbflanks')
ranges <- read_bed(bedfile)
flank_fourways(ranges)
}
