% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_genes_to_granges.R
\name{genes_to_granges}
\alias{genes_to_granges}
\title{Convert geneids into GRanges}
\usage{
genes_to_granges(geneids, txdb, plot = TRUE)
}
\arguments{
\item{geneids}{Entrezg vector}

\item{txdb}{\code{\link[GenomicFeatures]{TxDb-class}}}

\item{plot}{TRUE or FALSE}
}
\value{
\code{\link[GenomicRanges]{GRanges-class}}
}
\description{
Convert geneids into GRanges
}
\examples{
# Mouse - entrezgene
    geneids <- c('100009600', '99889', '99982')
    txdbname <- 'TxDb.Mmusculus.UCSC.mm10.knownGene'
    txdb <- utils::getFromNamespace(txdbname, txdbname)
    genes_to_granges(geneids, txdb)

# Mouse - ensembl
    ensdb <- EnsDb.Mmusculus.v98()
    geneids <- names(GenomicFeatures::genes(ensdb))
    genes_to_granges(geneids, ensdb)
    
    txdbname <- 'TxDb.Mmusculus.UCSC.mm10.ensGene'
    txdb <- utils::getFromNamespace(txdbname, txdbname)
    genes_to_granges(geneids, txdb)
    
# Human - entrezgene
    geneids <- c('1', '10', '100')
    txdbname <- 'TxDb.Hsapiens.UCSC.hg38.knownGene'
    txdb <- utils::getFromNamespace(txdbname, txdbname)
    genes_to_granges(geneids, txdb)
    
# Human - ensemblgene
    \dontrun{ # Takes a few minutes
    hub <- AnnotationHub::AnnotationHub()
    query(hub, c("Homo sapiens","EnsDb"))
    ensdb <- hub[["AH75011"]]
    geneids <- names(GenomicFeatures::genes(ensdb))[1:10]
    genes_to_granges(geneids, ensdb)
    }
}
\seealso{
\code{\link{annotate_granges}} (inverse operation)
}
