% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_genes_to_granges.R
\name{genes_to_granges}
\alias{genes_to_granges}
\title{Convert geneids into GRanges}
\usage{
genes_to_granges(geneids, db, plot = TRUE)
}
\arguments{
\item{geneids}{Entrezg vector}

\item{db}{\code{\link[GenomicFeatures]{TxDb-class}} or 
\code{\link[ensembldb]{EnsDb-class}}}

\item{plot}{TRUE or FALSE}
}
\value{
\code{\link[GenomicRanges]{GRanges-class}}
}
\description{
Convert geneids into GRanges
}
\examples{
# Mouse Entrez
    dbname <- 'TxDb.Mmusculus.UCSC.mm10.knownGene'
    db <- utils::getFromNamespace(dbname, dbname)
    geneids <- names(GenomicFeatures::genes(db))[1:100]
    genes_to_granges(geneids, db)

# Human Entrez
    dbname <- 'TxDb.Hsapiens.UCSC.hg38.knownGene'
    db <- utils::getFromNamespace(dbname, dbname)
    geneids <- names(GenomicFeatures::genes(db))[1:100]
    genes_to_granges(geneids, db)

# Mouse Ensembl
    db <- EnsDb.Mmusculus.v98()
    geneids <- names(GenomicFeatures::genes(db))[1:100]
    genes_to_granges(geneids, db)

# Human Ensembl
    db <- EnsDb.Hsapiens.v98()
    geneids <- names(GenomicFeatures::genes(db))[1:100]
    genes_to_granges(geneids, db)
}
\seealso{
\code{\link{annotate_granges}} (inverse operation)
}
