% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_xxx_to_granges.R
\name{genes_to_granges}
\alias{genes_to_granges}
\alias{genefile_to_granges}
\title{Convert geneids into GRanges}
\usage{
genes_to_granges(geneids, db, plot = TRUE)

genefile_to_granges(file, db, plot = TRUE)
}
\arguments{
\item{geneids}{Entrez Gene identifier vector}

\item{db}{\code{\link[GenomicFeatures]{TxDb-class}} or 
\code{\link[ensembldb]{EnsDb-class}}}

\item{plot}{TRUE or FALSE}

\item{file}{Entrez Gene identifier file (one per row)}
}
\value{
\code{\link[GenomicRanges]{GRanges-class}}
}
\description{
Convert geneids into GRanges
}
\examples{
 # Entrez
     dbname  <-'TxDb.Mmusculus.UCSC.mm10.knownGene'
     db      <- utils::getFromNamespace(dbname, dbname)
     genefile <- system.file('extdata/SRF.entrez', package='multicrispr')
     genefile_to_granges(genefile, db)
     geneids <- as.character(read.table(genefile)[[1]])
     genes_to_granges(geneids, db)

# Ensembl
     db <- EnsDb.Mmusculus.v98()
     genefile <- system.file('extdata/SRF.ensembl', package='multicrispr')
     genefile_to_granges(genefile, db)
     geneids <- as.character(read.table(genefile)[[1]])
     genes_to_granges(geneids, db)
     
}
\seealso{
\code{\link{annotate_granges}} (inverse operation)
}
