% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_bed_to_granges.R
\name{get_bsgenome}
\alias{get_bsgenome}
\title{Get BSgenome}
\usage{
get_bsgenome(gr)
}
\arguments{
\item{gr}{\code{\link[GenomicRanges]{GRanges-class}}}
}
\value{
BSgenome
}
\description{
Get BSgenome
}
\examples{
bedfile  <- system.file('extdata/SRF.bed', package='multicrispr')
gr <- bed_to_granges(bedfile, 'mm10')
get_bsgenome(gr)
}
