% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_read_assert.R
\name{get_bsgenome}
\alias{get_bsgenome}
\title{Get BSgenome}
\usage{
get_bsgenome(granges)
}
\arguments{
\item{granges}{GenomicRanges::GRanges}
}
\value{
BSgenome
}
\description{
Get BSgenome
}
\examples{
require(magrittr)
bedfile  <- system.file('extdata/SRF_sites.bed', package='crisprapex')
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
granges <- read_bed(bedfile, bsgenome) \%>\% flank_fourways()
granges \%>\% get_bsgenome()
}
