% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_add_offtargets.R
\name{has_been_indexed}
\alias{has_been_indexed}
\title{Has been indexed?}
\usage{
has_been_indexed(bsgenome, indexedgenomesdir = INDEXEDGENOMESDIR)
}
\arguments{
\item{bsgenome}{BSgenome}

\item{indexedgenomesdir}{directory with indexed genomes}
}
\description{
Has been indexed?
}
\examples{
bsgenome <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38
has_been_indexed(bsgenome)
}
