% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_alignment_offtargetanalysis.R
\name{index_genome}
\alias{index_genome}
\title{Index genome}
\usage{
index_genome(bsgenome, outdir = "~/.multicrispr/bowtie/genome")
}
\arguments{
\item{bsgenome}{\code{\link[BSgenome]{BSgenome-class}}}

\item{outdir}{string: output directory}
}
\value{
string: directory with indexed genome
}
\description{
Performs bowtie indexing of genome
}
\examples{
# index_genome(BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10)
# index_genome(BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38)
}
