% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_filter_specific.R
\name{index_genome}
\alias{index_genome}
\title{Index genome}
\usage{
index_genome(bsgenome, indexedgenome = default_indexedgenome(bsgenome))
}
\arguments{
\item{bsgenome}{\code{\link[BSgenome]{BSgenome-class}}}

\item{indexedgenome}{string: directory with bowtie-indexed genome}
}
\value{
string: directory with indexed genome
}
\description{
Bowtie index genome
}
\examples{
# index_genome(BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10)
# index_genome(BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38)
}
