% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_manipulate_ranges.R
\name{left_flank}
\alias{left_flank}
\alias{up_flank}
\alias{right_flank}
\alias{down_flank}
\alias{extend}
\title{Flank or Extend GRanges}
\usage{
left_flank(
  gr,
  leftstart = -200,
  leftend = -1,
  bsgenome = NULL,
  verbose = FALSE,
  plot = FALSE,
  ...
)

up_flank(
  gr,
  upstart = -200,
  upend = -1,
  bsgenome = NULL,
  verbose = FALSE,
  plot = FALSE,
  ...
)

right_flank(
  gr,
  rightstart = 1,
  rightend = 200,
  bsgenome = NULL,
  verbose = FALSE,
  plot = FALSE,
  ...
)

down_flank(
  gr,
  downstart = +1,
  downend = +200,
  bsgenome = NULL,
  verbose = FALSE,
  plot = FALSE,
  ...
)

extend(
  gr,
  leftstart = -22,
  rightend = 22,
  bsgenome = NULL,
  verbose = FALSE,
  plot = FALSE,
  ...
)
}
\arguments{
\item{gr}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{leftstart}{number: left start (relative to range start)}

\item{leftend}{number: left end   (relative to range start)}

\item{bsgenome}{NULL (default) or \code{\link[BSgenome]{BSgenome-class}}.
Required to update gr$seq if present.}

\item{verbose}{TRUE or FALSE (default)}

\item{plot}{TRUE or FALSE (default)}

\item{...}{passed to \code{\link{plot_intervals}}}

\item{rightstart}{number: right start (relative to range end)}

\item{rightend}{number: right end   (relative to range end)}
}
\value{
a \code{\link[GenomicRanges]{GRanges-class}}
}
\description{
Flank or Extend GRanges
}
\examples{
# Small GRanges
    bsgenome <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38
    gr <- GenomicRanges::GRanges(
            c(HBB = 'chr11:5227002-5227002', PRNP = 'chr20:4699500'), 
            strand = c('-', '+'), 
            seqinfo = BSgenome::seqinfo(bsgenome))
    gr \%>\% left_flank(-22, -1, plot = TRUE)
    gr \%>\% right_flank( 1, 22, plot = TRUE)
    gr \%>\%   up_flank(-22, -1, plot = TRUE)
    gr \%>\% down_flank(  1, 22, plot = TRUE)
    gr \%>\% extend(    -22, 22, plot = TRUE)

# Large GRanges
    require(magrittr)
    bedfile <- system.file('extdata/SRF.bed', package = 'multicrispr')
    gr <- bed_to_granges(bedfile, 'mm10', plot = FALSE)
    gr \%>\%   left_flank(-200,  -1, plot = TRUE)
    gr \%>\%  right_flank(   1, 200, plot = TRUE)
    gr \%>\%       extend(-200, 200, plot = TRUE)
}
\seealso{
\code{\link{straddle}} (single verb function encompassing all of 
         left_flank, right_flank, and extend) and \code{\link{double_flank}}.
}
