% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_flank_slop.R
\name{left_flank}
\alias{left_flank}
\title{Left flank}
\usage{
left_flank(granges, leftstart = -200, leftend = -1, verbose = TRUE)
}
\arguments{
\item{granges}{GenomicRanges::GRanges}

\item{leftstart}{left flank start (relative to range start)}

\item{leftend}{left flank end   (relative to range start)}

\item{verbose}{logical(1)}
}
\value{
GenomicRanges::GRanges
}
\description{
Left flank
}
\examples{
require(magrittr)
bedfile <- system.file('extdata/SRF_sites.bed', package = 'multicrispr')
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
granges <- read_bed(bedfile, bsgenome)
granges \%>\% head(3)
granges \%>\% head(3) \%>\% left_flank()
}
