% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_flank_slop.R
\name{left_flank}
\alias{left_flank}
\title{Left flank}
\usage{
left_flank(ranges, bsgenome, leftstart = -200, leftend = -1,
  verbose = TRUE)
}
\arguments{
\item{ranges}{data.table(chr, start, end, strand)}

\item{bsgenome}{BSgenome, e.g. BSgenome.Mmusculus.UCSC.mm10::Mmusculus}

\item{leftstart}{left flank start (relative to range start)}

\item{leftend}{left flank end   (relative to range start)}

\item{verbose}{logical(1)}
}
\value{
data.table
}
\description{
Left flank
}
\examples{
require(magrittr)
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
bedfile <- system.file('extdata/SRF_sites.bed', package = 'crisprapex')
ranges <- read_bed(bedfile)
ranges \%>\% head(3)
ranges \%>\% head(3) \%>\% left_flank(bsgenome)
}
