% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_read_flank_slop.R
\name{left_flank}
\alias{left_flank}
\title{Left flank}
\usage{
left_flank(ranges, startoffset = -200, endoffset = -1,
  verbose = TRUE)
}
\arguments{
\item{ranges}{data.table(chr, start, end)}

\item{startoffset}{flank start relative to range start}

\item{endoffset}{flank end   relative to range start}

\item{verbose}{logical(1)}
}
\value{
data.table
}
\description{
Left flank
}
\examples{
require(magrittr)
bedfile <- system.file('extdata/SRF_sites.bed', package = 'crisprapex')
ranges <- read_bed(bedfile)
ranges \%>\% head(1)
ranges \%>\% head(1) \%>\% left_flank()
}
