% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_flank_slop.R
\name{left_flank}
\alias{left_flank}
\title{Left flank}
\usage{
left_flank(gr, leftstart = -200, leftend = -1, plot = TRUE,
  verbose = TRUE)
}
\arguments{
\item{gr}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{leftstart}{number: flank start (relative to range start)}

\item{leftend}{number: flank end   (relative to range start)}

\item{plot}{logical(1)}

\item{verbose}{logical(1)}
}
\value{
a \code{\link[GenomicRanges]{GRanges-class}}
}
\description{
Left flank
}
\examples{
bedfile <- system.file('extdata/SRF.bed', package = 'multicrispr')
gr <- read_bed(bedfile, 'mm10', plot = FALSE)
left_flank(gr)
}
