% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_extend_flank.R
\name{left_flank}
\alias{left_flank}
\alias{right_flank}
\alias{extend}
\title{Flank or Extend GRanges}
\usage{
left_flank(
  gr,
  leftstart = -200,
  leftend = -1,
  bsgenome = NULL,
  plot = TRUE,
  verbose = TRUE
)

right_flank(
  gr,
  rightstart = 1,
  rightend = 200,
  bsgenome = NULL,
  plot = TRUE,
  verbose = TRUE
)

extend(
  gr,
  leftstart = -22,
  rightend = 22,
  bsgenome = NULL,
  plot = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{gr}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{leftstart}{number: left start (relative to range start)}

\item{leftend}{number: left end   (relative to range start)}

\item{bsgenome}{NULL (default) or \code{\link[BSgenome]{BSgenome-class}}.
Required to update gr$seq if present.}

\item{plot}{TRUE (default) or FALSE}

\item{verbose}{TRUE (default) or FALSE}

\item{rightstart}{number: right start (relative to range end)}

\item{rightend}{number: right end   (relative to range end)}
}
\value{
a \code{\link[GenomicRanges]{GRanges-class}}
}
\description{
Flank or Extend GRanges
}
\examples{
# SRF binding sites
    require(magrittr)
    bedfile <- system.file('extdata/SRF.bed', package = 'multicrispr')
    gr <- bed_to_granges(bedfile, 'mm10', plot = FALSE)
    gr \%>\% left_flank(-200,  -1)
    gr \%>\% right_flank(1, 200)
    gr \%>\% extend(-200, 200)
    
# HBB snp: sickle cell variant (T -> A)
    bsgenome <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38
    bsinfo <- BSgenome::seqinfo(bsgenome)
    gr <- GenomicRanges::GRanges(
            'chr11:5227002-5227002', strand = '-', seqinfo = bsinfo)
    (gr \%<>\% add_seq(bsgenome))
    gr \%>\% left_flank(-22, -1, bsgenome = bsgenome)
    gr \%>\% right_flank( 1, 22, bsgenome = bsgenome)
    gr \%>\% extend(-22, 22, bsgenome = bsgenome)

# PRNP snp: Kuru variant
   gr  <-  GenomicRanges::GRanges(
               'chr20:4699500', strand = '+', seqinfo = bsinfo)
   gr  \%<>\% multicrispr::add_inverse_strand()
   (gr \%<>\% multicrispr::add_seq(bsgenome))
   (extended <- multicrispr::extend(gr, bsgenome = bsgenome))
}
\seealso{
\code{\link{straddle}} (single verb function encompassing all of 
         left_flank, right_flank, and extend) and \code{\link{double_flank}}.
}
