% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_manipulate_ranges.R
\name{left_flank}
\alias{left_flank}
\alias{right_flank}
\alias{up_flank}
\alias{down_flank}
\alias{extend}
\title{Extend or Flank GRanges}
\usage{
left_flank(
  gr,
  start = -200,
  end = -1,
  bsgenome = NULL,
  verbose = FALSE,
  plot = FALSE,
  ...
)

right_flank(
  gr,
  start = 1,
  end = 200,
  bsgenome = NULL,
  verbose = FALSE,
  plot = FALSE,
  ...
)

up_flank(
  gr,
  start = -200,
  end = -1,
  bsgenome = NULL,
  verbose = FALSE,
  plot = FALSE,
  ...
)

down_flank(
  gr,
  start = +1,
  end = +200,
  bsgenome = NULL,
  verbose = FALSE,
  plot = FALSE,
  ...
)

extend(
  gr,
  start = -22,
  end = 22,
  bsgenome = NULL,
  verbose = FALSE,
  plot = FALSE,
  ...
)
}
\arguments{
\item{gr}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{start}{(pos or neg) number: relative start position (see details)}

\item{end}{(pos or neg) number: relative end position   (see details)}

\item{bsgenome}{NULL (default) or \code{\link[BSgenome]{BSgenome-class}}.
Required to update gr$seq if present.}

\item{verbose}{TRUE or FALSE (default)}

\item{plot}{TRUE or FALSE (default)}

\item{...}{passed to \code{\link{plot_intervals}}}
}
\value{
a \code{\link[GenomicRanges]{GRanges-class}}
}
\description{
Extend or left/right/up/down Flank GRanges
}
\details{
\code{      extend(gr,start,end)} = (grstart + start,   grend + end)
\code{  left_flank(gr,start,end)} = (grstart + start, grstart + end)
\code{ right_flank(gr,start,end)} = (  grend + start,   grend + end)
\code{ up_flank(gr,start,end)}    = (grstart + start, grstart + end)   '+'
                                    (  grend - end,   grend   - start) '-'
\code{ down_flank(gr,start,end)} = (  grend + start,   grend + end)    '+'
                                   (grstart - end,   grstart - start)  '-'
}
\examples{
bsgenome <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38
gr <- GenomicRanges::GRanges(
          c(locus1 = 'chr11:5227002', locus2 = 'chr11:4699500'), 
            strand = c('-', '+'), 
            seqinfo = BSgenome::seqinfo(bsgenome))
left_flank( gr,  -22, -1, plot = TRUE)
right_flank(gr,  +1, +22, plot = TRUE)
up_flank(   gr, -22,  -1, plot = TRUE)
down_flank( gr,  +1, +22, plot = TRUE)
extend(     gr, -22, +22, plot = TRUE)
}
