% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_filter_specific.R
\name{match_seqs}
\alias{match_seqs}
\title{Add match counts}
\usage{
match_seqs(
  seqs,
  indexdir,
  norc,
  outfile = paste0(default_outdir(), "/match_", basename(indexdir), ".txt"),
  count_vars = c("MM0", "MM1", "MM2"),
  verbose = TRUE
)
}
\arguments{
\item{seqs}{character vector: sequences to match against indexed ref}

\item{indexdir}{string: dir containing indexed reference.
This can be an indexed genome( \code{\link{index_genome}}
It can also be indexed targets (\code{\link{index_targets}})}

\item{norc}{TRUE or FALSE: whether to run bowtie also with revcompls
Generally TRUE for genome and FALSE for target matches, 
because target ranges generally include both strands.}

\item{outfile}{string: file where to output bowtie results}

\item{count_vars}{character(3): how to name count variables}

\item{verbose}{TRUE (default) or FALSE}
}
\description{
Count matches to indexed target/genome and add to GRanges
}
\details{
\code{match_seqs} matches sequences against (indexed) target and genome
\code{add_crispr_matches} expands iupac amgiguities in the pam sequence, 
matches all resulting sequences against (indexes) target and genome, 
adds match counts to GRanges object, and then returns it
}
\examples{
# TFBS example
#-------------
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
bedfile  <- system.file('extdata/SRF.bed', package = 'multicrispr')
targets <- extend(bed_to_granges(bedfile, genome = 'mm10'))
indexedtargets <- index_targets(targets, bsgenome, indexedtargets)
spacers <- find_spacers(targets, bsgenome)
crisprseqs <- unique(paste0(spacers$crisprspacer, spacers$crisprpam))
match_seqs(crisprseqs, indexedtargets, norc=FALSE)
}
\seealso{
\code{\link{index_genome}}, \code{\link{index_targets}}
}
