% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_filter_specific.R
\name{match_seqs}
\alias{match_seqs}
\title{Match spacer sequences}
\usage{
match_seqs(
  seqs,
  indexdir,
  norc,
  mismatches = 2,
  outdir = OUTDIR,
  verbose = TRUE
)
}
\arguments{
\item{seqs}{character vector: sequences to match against indexed ref}

\item{indexdir}{string: dir containing indexed reference.
This can be an indexed genome( \code{\link{index_genome}}
It can also be indexed targets (\code{\link{index_targets}})}

\item{norc}{TRUE or FALSE: whether to run bowtie also with revcompls
Generally TRUE for genome and FALSE for target matches, 
because target ranges generally include both strands.}

\item{mismatches}{max number of mismatches to consider}

\item{outdir}{string: multicrispr output directory}

\item{verbose}{TRUE (default) or FALSE}
}
\description{
Count matches to indexed target/genome
}
\examples{
# TFBS example
#-------------
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
bedfile  <- system.file('extdata/SRF.bed', package = 'multicrispr')
targets <- extend(bed_to_granges(bedfile, genome = 'mm10'))
indexdir <- index_targets(targets, bsgenome)
spacers <- find_spacers(targets, bsgenome)
seqs <- unique(paste0(spacers$crisprspacer, spacers$crisprpam))
match_seqs(seqs, indexdir, norc=FALSE)
match_seqs(seqs, indexdir, norc=FALSE, mismatches=3)
}
\seealso{
\code{\link{index_genome}}, \code{\link{index_targets}}
}
