% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_filter_specific.R
\name{match_spacers}
\alias{match_spacers}
\title{Match spacers}
\usage{
match_spacers(
  spacers,
  indexdir,
  norc,
  mismatches = 2,
  outdir = OUTDIR,
  pam = "NGG",
  verbose = TRUE
)
}
\arguments{
\item{spacers}{spacer \code{\link[GenomicRanges]{GRanges-class}}}

\item{indexdir}{string: dir containing indexed reference.
This can be an indexed genome( \code{\link{index_genome}}
It can also be indexed targets (\code{\link{index_targets}})}

\item{norc}{TRUE or FALSE: whether to run bowtie also with revcompls
Generally TRUE for genome and FALSE for target matches, 
because target ranges generally include both strands.}

\item{mismatches}{number (default 2): max number of mismatches to consider}

\item{outdir}{string: file where to output bowtie results}

\item{pam}{string (default 'NGG') pam pattern to expand}

\item{verbose}{TRUE (default) or FALSE}
}
\value{
data.table
}
\description{
Count matches to indexed target/genome and add to GRanges
}
\details{
Expands iupac amgiguities in the pam sequence.
Matches all resulting sequences against (indexes) target and genome.
Adds match counts to GRanges object, and then returns it.
}
\examples{
# TFBS example
#-------------
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
bedfile  <- system.file('extdata/SRF.bed', package = 'multicrispr')
targets <- extend(bed_to_granges(bedfile, genome = 'mm10'))
indexdir <- index_targets(targets, bsgenome)
spacers <- find_spacers(targets, bsgenome)
match_spacers(spacers, indexdir, norc=FALSE, mismatches = 1)
}
\seealso{
\code{\link{index_genome}}, \code{\link{index_targets}}
}
