% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plot.R
\name{plot_intervals}
\alias{plot_intervals}
\title{Interval plot GRanges}
\usage{
plot_intervals(
  gr,
  color_var = if ("color" \%in\% names(mcols(gr))) "color" else "seqnames",
  title = NULL
)
}
\arguments{
\item{gr}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{color_var}{string: mcol mapped to color.}

\item{title}{plot title}
}
\value{
ggplot object
}
\description{
Interval plot GRanges
}
\examples{
# SRF sites
    require(magrittr)
    bedfile <-  system.file('extdata/SRF.bed',  package = 'multicrispr')
    sites   <- bed_to_granges(bedfile, 'mm10', plot = FALSE)
    plot_intervals(sites)
    
    flanks  <- left_flank(sites)
    sites$color <- 'sites'
    flanks$color <- 'flanks'
    plot_intervals(c(sites, flanks))
    
# PE targets
    bs <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38  
    sites <- GenomicRanges::GRanges(
               seqnames= c(PRNP = 'chr20:4699600',
                           HBB  = 'chr11:5227002',
                           HEXA = 'chr15:72346580-72346583',
                           CFTR = 'chr7:117559593-117559595'),
               strand   = c(PRNP = '+', HBB = '-', HEXA = '-', CFTR = '+'), 
               seqinfo  = BSgenome::seqinfo(bs))
    sites$color <- names(sites)
    plot_intervals(sites)
}
\seealso{
\code{\link{plot_karyogram}}
}
