% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_plot.R
\name{plot_karyogram}
\alias{plot_karyogram}
\title{Karyo/Interval Plot GRanges(List)}
\usage{
plot_karyogram(grlist, title = unique(GenomeInfoDb::genome(grlist)))
}
\arguments{
\item{grlist}{\code{\link[GenomicRanges]{GRanges-class}} or 
\code{\link[GenomicRanges]{GRangesList-class}}}

\item{title}{plot title}
}
\value{
list (plot_karyogram) or ggplot (plot_intervals)
}
\description{
Karyo/Interval Plot GRanges(List)
}
\examples{
# Plot GRanges
  bedfile <-  system.file('extdata/SRF.bed',  package = 'multicrispr')
  gr <- bed_to_granges(bedfile, 'mm10', plot = FALSE)
  plot_karyogram(gr)
  
# Plot GRangesList
  flanks  <- left_flank(gr)
  grlist <- GenomicRanges::GRangesList(sites = gr, flanks = flanks)
  plot_karyogram(grlist)
}
\seealso{
\code{\link{plot_intervals}}
}
