% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_plot.R
\name{plot_karyogram}
\alias{plot_karyogram}
\alias{plot_intervals}
\title{Karyo/Interval Plot GRanges(List)}
\usage{
plot_karyogram(grangeslist, title = unique(GenomeInfoDb::genome(grangeslist)))

plot_intervals(grangeslist, title = NULL)
}
\arguments{
\item{grangeslist}{\code{\link[GenomicRanges]{GRanges-class}} or 
\code{\link[GenomicRanges]{GRangesList-class}}}

\item{title}{plot title}
}
\value{
list (plot_karyogram) or ggplot (plot_intervals)
}
\description{
Karyo/Interval Plot GRanges(List)
}
\examples{
# Plot GRanges
  bedfile <-  system.file('extdata/SRF.bed',  package = 'multicrispr')
  gr <- bed_to_granges(bedfile, 'mm10', plot = FALSE)
  plot_karyogram(gr)
  plot_intervals(gr)
  
# Plot GRangesList
  flanks  <- left_flank(gr)
  grangeslist <- GenomicRanges::GRangesList(sites = gr, flanks = flanks)
  plot_karyogram(grangeslist)
  plot_intervals(grangeslist)
}
\seealso{
\code{\link[karyoploteR]{plotKaryotype}}, around which this 
          function wraps
}
