% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_find_cas9sites.R
\name{range2seq}
\alias{range2seq}
\title{Convert range to sequence}
\usage{
range2seq(chr, start, end, strand, bsgenome)
}
\arguments{
\item{chr}{chromosome vector}

\item{start}{start position vector}

\item{end}{vector end position vector}

\item{strand}{'+|-' vector}

\item{bsgenome}{BSgenome object}
}
\description{
Convert range to sequence
}
\examples{
range2seq(
    chr      = c('chr2', 'chr1', 'chr17'),
    start    = c(83972889, 85603908, 70963622),
    end      = c(83972905, 85603924, 70963638),
    strand   = c('+', '-', '-'),
    bsgenome = BSgenome.Mmusculus.UCSC.mm10::Mmusculus)
}
