% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_read_assert.R
\name{read_bed}
\alias{read_bed}
\title{Read bedfile into GRanges}
\usage{
read_bed(bedfile, bsgenome, zero_based = TRUE, rm_duplicates = TRUE,
  plot = TRUE, verbose = TRUE)
}
\arguments{
\item{bedfile}{file path}

\item{bsgenome}{BSgenome, e.g. BSgenome.Mmusculus.UCSC.mm10::Mmusculus}

\item{zero_based}{logical(1): whether bedfile GRanges are 0-based}

\item{rm_duplicates}{logical(1)}

\item{plot}{logical(1)}

\item{verbose}{logical(1)}
}
\value{
\code{\link[GenomicRanges]{GRanges-class}}
}
\description{
Read bedfile into GRanges
}
\note{
By convention BED files are 0-based. GRanges are always 1-based.
      A good discussion on these two alternative codings is given 
      by Obi Griffith on Biostars: https://www.biostars.org/p/84686/
}
\examples{
bedfile  <- system.file('extdata/SRF.bed', package = 'multicrispr')
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
(gr <- read_bed(bedfile, bsgenome))
}
