% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_read_flank_slop.R
\name{read_bed}
\alias{read_bed}
\title{Read bedfile as data.table}
\usage{
read_bed(bedfile, verbose = TRUE, rm_duplicates = TRUE)
}
\arguments{
\item{bedfile}{file path}

\item{verbose}{logical(1)}

\item{rm_duplicates}{logical(1)}
<<<<<<< HEAD
}
\value{
data.table(chr, start, end, strand)
=======
>>>>>>> 98ea718d0db705cc5f915d0635d5c327da17e58a
}
\description{
Read bedfile as data.table
}
\examples{
bedfile <- system.file('extdata/SRF_sites.bed', package = 'cas9tbflanks')
ranges <- read_bed(bedfile)
}
