% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_read_assert.R
\name{read_bed}
\alias{read_bed}
\title{Read bedfile as data.table}
\usage{
read_bed(bedfile, bsgenome, verbose = TRUE, rm_duplicates = TRUE)
}
\arguments{
\item{bedfile}{file path}

\item{bsgenome}{BSgenome, e.g. BSgenome.Mmusculus.UCSC.mm10::Mmusculus}

\item{verbose}{logical(1)}

\item{rm_duplicates}{logical(1)}
}
\value{
data.table(chr, start, end, strand)
}
\description{
Read bedfile as data.table
}
\examples{
bedfile <- system.file('extdata/SRF_sites.bed', package = 'multicrispr')
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
read_bed(bedfile, bsgenome)
}
