% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_read.R
\name{read_bed}
\alias{read_bed}
\title{Read bedfile into GRanges}
\usage{
read_bed(bedfile, genome, plot = TRUE, verbose = TRUE)
}
\arguments{
\item{bedfile}{file path}

\item{genome}{character: e.g. 'mm10'}

\item{plot}{logical(1)}

\item{verbose}{logical(1)}
}
\value{
\code{\link[GenomicRanges]{GRanges-class}}
}
\description{
Read bedfile into granges and output graphical and textual summaries
}
\examples{
bedfile  <- system.file('extdata/SRF.bed', package = 'multicrispr')
read_bed(bedfile, 'mm10')
}
\seealso{
\code{rtracklayer::import.bed} (documented in 
\code{\link[rtracklayer]{BEDFile-class}}), around which this function wraps.
}
