% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_flank_slop.R
\name{right_flank}
\alias{right_flank}
\title{Right flank}
\usage{
right_flank(gr, rightstart = 1, rightend = 200, plot = TRUE,
  verbose = TRUE)
}
\arguments{
\item{gr}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{rightstart}{number: flank start (relative to range end)}

\item{rightend}{number: flank end   (relative to range end)}

\item{plot}{logical(1)}

\item{verbose}{logical(1)}
}
\value{
\code{\link[GenomicRanges]{GRanges-class}}
}
\description{
Right flank
}
\examples{
bedfile <- system.file('extdata/SRF.bed', package = 'multicrispr')
gr <- read_bed(bedfile, 'mm10', plot = FALSE)
right_flank(gr)
}
