% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_flank_slop.R
\name{right_flank}
\alias{right_flank}
\title{Right flank}
\usage{
right_flank(granges, rightstart = 1, rightend = 200, verbose = TRUE)
}
\arguments{
\item{granges}{data.table(chr, start, end)}

\item{rightstart}{flank start relative to range start}

\item{verbose}{logical(1)}

\item{leftend}{flank end   relative to range start}
}
\value{
GenomicRanges::GRanges
}
\description{
Right flank
}
\examples{
require(magrittr)
bedfile <- system.file('extdata/SRF_sites.bed', package = 'multicrispr')
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
granges <- read_bed(bedfile, bsgenome)
granges \%>\% head(1)
granges \%>\% head(1) \%>\% right_flank()
}
