% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_find_cas9s.R
\name{rm_taboo_cas9s}
\alias{rm_taboo_cas9s}
\title{Rm taboo cas9s}
\usage{
rm_taboo_cas9s(targetcas9s, taboocas9s, verbose = TRUE)
}
\arguments{
\item{targetcas9s}{data.table(cas9seq, ...)}

\item{taboocas9s}{data.table(cas9seq, ...)}

\item{verbose}{logical(1)}
}
\value{
subset of taboocas9s
}
\description{
Rm taboo cas9s
}
\examples{
require(magrittr)
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
bedfile <- system.file('extdata/SRF_sites.bed', package='crisprapex')
tbranges <- read_bed(bedfile)
flankcas9s  <- tbranges \%>\% flank_fourways() \%>\% findcas9s(bsgenome)
centercas9s <- tbranges \%>\% slop_fourways()  \%>\% findcas9s(bsgenome)
flankcas9s \%>\% rm_taboo_cas9s(centercas9s)

}
