% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_score_cas9s.R
\name{score_cas9ranges}
\alias{score_cas9ranges}
\title{Score cas9ranges}
\usage{
score_cas9ranges(cas9ranges, ruleset = 1, verbose = TRUE)
}
\arguments{
\item{cas9ranges}{GenomicRanges::GRanges}

\item{ruleset}{1 (default) or 2 (only if python module 
github/MicrosoftResearch/azimuth is installed)}

\item{verbose}{logical(1)}
}
\value{
numeric vector
}
\description{
Score cas9ranges using ruleset1 (Doench 2014) or ruleset2 (Doench 2016)
}
\details{
For ruleset1, the CRISPRseek implementation is used.
For ruleset2, azimuth (github/MicrosoftResearch/azimuth) is used.
}
\examples{

# Get cas9ranges
    require(magrittr)
    bedfile <- system.file('extdata/SRF_sites.bed', package = 'crisprapex')
    bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
    targetranges <- read_bed(bedfile, bsgenome) \%>\% 
                    flank_fourways() \%>\% 
                    find_cas9s()
# Score
    cas9ranges[1:10] \%>\% score_cas9ranges()
    cas9ranges[1:10] \%>\% contextseqs() \%>\% score_contextseqs()

}
\references{
Doench 2014, Rational design of highly active sgRNAs for 
CRISPR-Cas9-mediated gene inactivation. Nature Biotechnology,
doi: 10.1038/nbt.3026

Doench 2016, Optimized sgRNA design to maximize activity and minimize 
off-target effects of CRISPR-Cas9. Nature Biotechnology, 
doi: 10.1038/nbt.3437
}
