% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_score_cas9s.R
\name{score_cas9ranges}
\alias{score_cas9ranges}
\title{Score cas9ranges}
\usage{
score_cas9ranges(chr, start, end, strand, bsgenome, verbose = TRUE)
}
\arguments{
\item{chr}{character vector: chromosome values}

\item{start}{numeric vector: start positions}

\item{end}{numeric vector: end positions}

\item{strand}{character vector: '+' or '-' values}

\item{bsgenome}{BSGenome, e.g. BSgenome.Mmusculus.UCSC.mm10::Mmusculus}

\item{verbose}{logical(1)}
}
\value{
numeric vector
}
\description{
Score cas9ranges
}
\examples{
do_run <-   assertive.reflection::is_unix() &
            reticulate::py_module_available('azimuth')
if (do_run){
    require(magrittr)
    bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
    bedfile <- system.file('extdata/SRF_sites.bed', package = 'crisprapex')
    cas9dt <-   read_bed(bedfile)        \%>\% 
                head(3)                  \%>\% 
                slop_fourways(bsgenome)  \%>\% 
                find_cas9s(bsgenome)
    cas9dt [ , score_cas9ranges(chr, cas9start, cas9end, strand, bsgenome) ]
}
}
