% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_score_cas9s.R
\name{score_contextseqs}
\alias{score_contextseqs}
\alias{score_cas9ranges}
\title{Score cas9ranges}
\usage{
score_contextseqs(contextseqs, ruleset = 1, verbose = TRUE,
  python = NULL, virtualenv = NULL, condaenv = NULL)

score_cas9ranges(cas9ranges, ruleset = 1, verbose = TRUE,
  python = NULL, virtualenv = NULL, condaenv = NULL)
}
\arguments{
\item{contextseqs}{character vector with 4-23-3 contextseqs}

\item{ruleset}{1 (default) or 2 (only if python module 
github/MicrosoftResearch/azimuth is installed)}

\item{verbose}{logical(1)}

\item{python}{NULL (ruleset=1) or path to a python binary (ruleset=2). See details.}

\item{virtualenv}{NULL (ruleset=1) or directory containing python virtualenv (ruleset=2). See details.}

\item{condaenv}{NULL (ruleset=1) or name of condaenv (ruleset=2). See details.}

\item{cas9ranges}{\code{\link[GenomicRanges]{GRanges-class}}}
}
\value{
numeric vector
}
\description{
Score cas9ranges using ruleset1 (Doench 2014) or ruleset2 (Doench 2016)
}
\details{
ruleset1 is readily available. ruleset2 is accessible after installing 
the python module [azimuth](https://github.com/MicrosoftResearch/Azimuth), 
and specifying a value for either 'python' (python binary path), 
'virtualenv' (python virtual environment dir) or 
'condaenv' (python conda environment).
}
\examples{

# Get cas9ranges
    require(magrittr)
    bedfile <- system.file('extdata/SRF.bed', package = 'multicrispr')
    cas9ranges <- read_bed(bedfile, 'mm10') \%>\% 
                  double_flank() \%>\% 
                  find_cas9s()
# Score
    cas9ranges[1:3] \%>\% score_cas9ranges()
    cas9ranges[1:3] \%>\% contextseqs() \%>\% score_contextseqs()

}
\references{
Doench 2014, Rational design of highly active sgRNAs for 
CRISPR-Cas9-mediated gene inactivation. Nature Biotechnology,
doi: 10.1038/nbt.3026

Doench 2016, Optimized sgRNA design to maximize activity and minimize 
off-target effects of CRISPR-Cas9. Nature Biotechnology, 
doi: 10.1038/nbt.3437
}
