% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_score_cas9s.R
\name{score_contextseqs}
\alias{score_contextseqs}
\title{Score contextualized cas9seqs}
\usage{
score_contextseqs(contextseqs, ruleset = 1, verbose = TRUE)
}
\arguments{
\item{contextseqs}{30-char string vector: 4-23-3 (flank-cas9-flank)}

\item{ruleset}{1 (Doench 2014 - CRISPRseek) or 
2 (Doench 2016 - github/MicrosoftResearch/azimuth)}

\item{verbose}{logical(1)}
}
\value{
numeric vector
}
\description{
Score contextualized cas9seqs using ruleset1 (Doench2014) or
ruleset2 (Doench2016)
}
\examples{
require(magrittr)
contextseqs <- c('TGCCCTTATATTGTCTCCAGCAGAAGGTGT',
                 'TGCCCTTATATTGTCTCCAGCAGAAGGTGT',
                 'CCAAATATTGTCAAGTTGACAACCAGGAAT')
contextseqs \%>\% score_contextseqs()
}
\references{
Doench 2014, Rational design of highly active sgRNAs for 
CRISPR-Cas9-mediated gene inactivation. Nature Biotechnology,
doi: 10.1038/nbt.3026

Doench 2016, Optimized sgRNA design to maximize activity and minimize 
off-target effects of CRISPR-Cas9. Nature Biotechnology, 
doi: 10.1038/nbt.3437
}
