% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_find_cas9s.R
\name{seqs}
\alias{seqs}
\title{Convert GRanges into Sequences}
\usage{
seqs(granges)
}
\arguments{
\item{granges}{\code{\link[GenomicRanges]{GRanges-class}}}
}
\value{
character vector
}
\description{
Convert GRanges into Sequences
}
\examples{
bedfile  <- system.file('extdata/SRF.bed', package = 'multicrispr')
granges <- bed_to_granges(bedfile, 'mm10')
seqs(granges)
}
