% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_find_cas9s.R
\name{seqs}
\alias{seqs}
\title{Get sequence values}
\usage{
seqs(granges)
}
\arguments{
\item{granges}{GenomicRanges::GRanges}
}
\value{
DNAStringSet (get) or GRanges (set)
}
\description{
Get sequence values
}
\examples{
bedfile  <- system.file('extdata/SRF_sites.bed', package = 'multicrispr')
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
granges <- read_bed(bedfile, bsgenome)
seqs(granges)
}
