% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_read_flank_slop.R
\name{slop}
\alias{slop}
\title{Slop (i.e. extend left/right)}
\usage{
slop(ranges, bsgenome, startoffset = -22, endoffset = 22,
  verbose = TRUE)
}
\arguments{
\item{ranges}{data.table(chr, start, end)}

\item{bsgenome}{BSgenome, e.g. BSgenome.Mmusculus.UCSC.mm10::Mmusculus}

\item{startoffset}{flank start relative to range start}

\item{endoffset}{flank end   relative to range start}

\item{verbose}{logical(1)}
}
\value{
data.table
}
\description{
Slop (i.e. extend left/right)
}
\examples{
require(magrittr)
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
bedfile <- system.file('extdata/SRF_sites.bed', package = 'crisprapex')
ranges <- read_bed(bedfile)
ranges \%>\% head(1)
ranges \%>\% head(1) \%>\% slop(bsgenome)
}
