% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_read_flank_slop.R
\name{slop}
\alias{slop}
\title{Slop (i.e. extend left/right)}
\usage{
slop(ranges, startoffset = -22, endoffset = 22, verbose = TRUE)
}
\arguments{
\item{ranges}{data.table(chr, start, end)}

\item{startoffset}{flank start relative to range start}

\item{endoffset}{flank end   relative to range start}

\item{verbose}{logical(1)}
}
\value{
data.table
}
\description{
Slop (i.e. extend left/right)
}
\examples{
require(magrittr)
bedfile <- system.file('extdata/SRF_sites.bed', package = 'cas9tbflanks')
ranges <- read_bed(bedfile)
ranges \%>\% head(1)
ranges \%>\% head(1) \%>\% slop()
}
