% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_flank_slop.R
\name{slop}
\alias{slop}
\title{Slop (i.e. extend left/right)}
\usage{
slop(gr, leftstart = -22, rightend = 22, plot = TRUE,
  verbose = TRUE)
}
\arguments{
\item{gr}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{leftstart}{number: flank start (relative to range start)}

\item{rightend}{number: flank end   (relative to range end)}

\item{plot}{logical(1)}

\item{verbose}{logical(1)}
}
\value{
\code{\link[GenomicRanges]{GRanges-class}}
}
\description{
Slop (i.e. extend left/right)
}
\examples{
bedfile <- system.file('extdata/SRF.bed', package = 'multicrispr')
gr <- read_bed(bedfile, 'mm10', plot = FALSE)
slop(gr)
}
