% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_flank_slop.R
\name{slop_fourways}
\alias{slop_fourways}
\title{Slop ranges for both strands, merging overlaps}
\usage{
slop_fourways(ranges, bsgenome, leftstart = -22, rightend = 22,
  verbose = TRUE)
}
\arguments{
\item{ranges}{data.table(chr, start, end, strand)}

\item{bsgenome}{BSgenome, e.g. BSgenome.Mmusculus.UCSC.mm10::Mmusculus}

\item{leftstart}{numeric(1)}

\item{rightend}{numeric(1)}

\item{verbose}{logical(1)}
}
\value{
data.table(chr, start, end, strand)
}
\description{
Slop ranges for both strands, merging overlaps
}
\examples{
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
bedfile <- system.file('extdata/SRF_sites.bed', package = 'crisprapex')
ranges <- read_bed(bedfile)
slop_fourways(ranges, bsgenome)
}
