% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_flank_slop.R
\name{slop_fourways}
\alias{slop_fourways}
\title{Slop granges for both strands, merging overlaps}
\usage{
slop_fourways(granges, leftstart = -22, rightend = 22,
  verbose = TRUE)
}
\arguments{
\item{granges}{GenomicRanges::GRanges}

\item{leftstart}{numeric(1)}

\item{rightend}{numeric(1)}

\item{verbose}{logical(1)}
}
\value{
GenomicRanges::GRanges
}
\description{
Slop granges for both strands, merging overlaps
}
\examples{
require(magrittr)
bedfile <- system.file('extdata/SRF_sites.bed', package = 'multicrispr')
bsgenome <- BSgenome.Mmusculus.UCSC.mm10::Mmusculus
granges <- read_bed(bedfile, bsgenome)
granges
granges \%>\% slop_fourways()
}
