% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_read_flank_slop.R
\name{slop_fourways}
\alias{slop_fourways}
\title{Slop ranges for both strands, merging overlaps}
\usage{
slop_fourways(ranges, startoffset = -22, endoffset = 22,
  verbose = TRUE)
}
\arguments{
\item{ranges}{data.table(chr, start, end, strand)}

\item{startoffset}{numeric(1)}

\item{endoffset}{numeric(1)}

\item{verbose}{logical(1)}
}
\value{
data.table(chr, start, end, strand)
}
\description{
Slop ranges for both strands, merging overlaps
}
\examples{
bedfile <- system.file('extdata/SRF_sites.bed', package = 'cas9tbflanks')
ranges <- read_bed(bedfile)
slop_fourways(ranges)
}
