% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_straddle.R
\name{left_flank}
\alias{left_flank}
\alias{right_flank}
\alias{extend}
\alias{straddle}
\title{Extend or flank GRanges}
\usage{
left_flank(gr, leftstart = -200, leftend = -1, plot = TRUE,
  verbose = TRUE)

right_flank(gr, rightstart = 1, rightend = 200, plot = TRUE,
  verbose = TRUE)

extend(gr, leftstart = -22, rightend = 22, plot = TRUE,
  verbose = TRUE)

straddle(gr, leftstart = NULL, leftend = NULL, rightstart = NULL,
  rightend = NULL, plot = TRUE, verbose = TRUE)
}
\arguments{
\item{gr}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{leftstart}{number: left start (relative to range start)}

\item{leftend}{number: left end   (relative to range start)}

\item{plot}{logical(1)}

\item{verbose}{logical(1)}

\item{rightstart}{number: right start (relative to range end)}

\item{rightend}{number: right end   (relative to range end)}
}
\value{
a \code{\link[GenomicRanges]{GRanges-class}}
}
\description{
Extend or flank GRanges
}
\examples{
# Read ranges
    bedfile <- system.file('extdata/SRF.bed', package = 'multicrispr')
    gr <- bed_to_granges(bedfile, 'mm10', plot = FALSE)

# Flank/Extend
    left_flank( gr, -200,  -1)
    right_flank(gr,    1, 200)
    extend(     gr, -200, 200)
    
# Single-verb alternative
    straddle(gr, leftstart=-200, leftend=-1)    # left flank
    straddle(gr, rightstart=1,   rightend=200)  # right flank
    straddle(gr, leftstart=-200, rightend=200)  # extend
}
