% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_manipulate_ranges.R
\name{up_flank}
\alias{up_flank}
\alias{down_flank}
\alias{extend}
\title{Extend or Flank GRanges}
\usage{
up_flank(
  gr,
  start = -200,
  end = -1,
  stranded = TRUE,
  bsgenome = NULL,
  verbose = FALSE,
  plot = FALSE,
  ...
)

down_flank(
  gr,
  start = 1,
  end = 200,
  stranded = TRUE,
  bsgenome = NULL,
  verbose = FALSE,
  plot = FALSE,
  ...
)

extend(
  gr,
  start = -22,
  end = 22,
  stranded = TRUE,
  bsgenome = NULL,
  verbose = FALSE,
  plot = FALSE,
  ...
)
}
\arguments{
\item{gr}{\code{\link[GenomicRanges]{GRanges-class}}}

\item{start}{(pos or neg) number: relative start position (see details)}

\item{end}{(pos or neg) number: relative end position   (see details)}

\item{stranded}{TRUE (default) or FALSE: consider strand information?}

\item{bsgenome}{NULL (default) or \code{\link[BSgenome]{BSgenome-class}}.
Required to update gr$seq if present.}

\item{verbose}{TRUE or FALSE (default)}

\item{plot}{TRUE or FALSE (default)}

\item{...}{passed to \code{\link{plot_intervals}}}
}
\value{
a \code{\link[GenomicRanges]{GRanges-class}}
}
\description{
Returns extensions, upstream flanks, or downstream flanks
}
\details{
\code{up_flank}   returns upstream flanks, in relation to start(gr).
\code{down_flank} returns downstream flanks, in relation to end(gr).
\code{extend}     returns extensions, in relation to start(gr) and end(gr)
}
\examples{
# PE example
#-----------
    require(magrittr)
    bsgenome <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38  
    gr <- GenomicRanges::GRanges(
              seqnames = c(PRNP = 'chr20:4699600',             # snp
                           HBB  = 'chr11:5227002',             # snp
                           HEXA = 'chr15:72346580-72346583',   # del
                           CFTR = 'chr7:117559593-117559595'), # ins
              strand   = c(PRNP = '+', HBB = '-', HEXA = '-', CFTR = '+'), 
              seqinfo  = BSgenome::seqinfo(bsgenome))
      up_flank(gr, -22,  -1, plot = TRUE)
      up_flank(gr, -22,  -1, plot = TRUE, stranded = FALSE)
    down_flank(gr,  +1, +22, plot = TRUE)
    down_flank(gr,  +1, +22, plot = TRUE, stranded = FALSE)
        extend(gr, -10, +20, plot = TRUE)
        extend(gr, -10, +20, plot = TRUE, stranded = FALSE)
}
